LoW: Initialize units in start events
previously this was done in side 1 turn 1 events because start events
wasn't synced.

After the fix for 21933 players now loose when they have no units after
start events, so we need to do this in a start event.

replacing check_victory with check_end_level after start events in
play_controller.cpp line 572 doesn't work because we still have the
check_victory in playmp_controller when waiting
for remote packages, that check_victory is needed in order to sync the
behavior of ai turns which calls check_victory after every action on the
local client.


Former-commit-id: 64c0b025e4c471e4672c35a0bbbb02a2ef7363da
Former-commit-id: 437ce38260fbe566125cf7dd5ef37e86285f0c22
Former-commit-id: bea39aa69dd123010eb0b2c59c2980e169d28b52
Former-commit-id: c5577a046419ec6729f57300b2e31bcb42a8b71c [formerly 48d84f2b1245d93d01fd3f1d453e823d800836a9]
Former-commit-id: 6d4f1f88f790d5abe42737e51f67901a8fb39663