Goto Micro AI: prevent potential conflict between Goto MAIs
If several Goto MAIs are used on the same side and both have
unique_goals=yes set, they previously might have interfered with each
other, one preventing the other from working correctly. Now the “goal
taken” information is saved such that it only applies to the CA by
which it was set.


Former-commit-id: a2aad2a7058cea31c8031026be74339187411e4e
Former-commit-id: d5425fcbcd98c8c5ca4b9d6e273cae8b31e113e2
Former-commit-id: 13b802e191bb31c2023c58a1ba7a4e9d7b69161a
Former-commit-id: 79dc7b7bdd798f84d7d8cab5d82fd5dd0fc8efc1 [formerly fda6bcaca21aeb37158db1b31bb578ebd1b5616a]
Former-commit-id: 08517e6565540dea3eb430f803f289af9d5b0274