Goto Micro AI: prevent potential conflict between Goto MAIs
If several Goto MAIs are used on the same side and both have
unique_goals=yes set, they previously might have interfered with each
other, one preventing the other from working correctly. Now the “goal
taken” information is saved such that it only applies to the CA by
which it was set.


Former-commit-id: e4ee4d9a451e8015013603fec3c2512d1cb1a35b
Former-commit-id: d7f18ab6ac8c5bf49df0b141438b0294d3b95111
Former-commit-id: 27cd39d6c15279a8d4aa0d9a0ac5b4adcbea281f
Former-commit-id: f38682969bfdcfaaf9ca5f9254f502ffe7c1f652 [formerly 7fad938cae2b09b30714f9a577e30cc22189aa5d]
Former-commit-id: 0809e13ee585351c30874ca7d06bf8f56bbf94f2