partly revert 301b2513aa3d [formerly d435fa477de8] [formerly 89c237804bcd [formerly 984e68f39362]] [formerly b0241676fec3] [formerly 79c3751fafd5] [formerly 8d42175cb748]
this should restore the behaviour that players don't loose if they have
no units left after the start event. Because it wasn't tested enough.

Note, that in most cases we don't need the check_victory in
playmp_controller::play_network_turn because we have a check_victory()
in the end in run_in_sycned_command


Former-commit-id: 18f255d633fdcec16580308602fe0acd3539c36d
Former-commit-id: d2e2bf1f7622a0cf6df66a4bb516b8874df3be38
Former-commit-id: 8dc427f3e3b8bb37802b7207d48af5a0ee9b9067
Former-commit-id: 8e8a5f24eb1627a72b0b532e9196ef13198901c1 [formerly cbd430fe1574f71cf942664ece59c9e0ce89592e]
Former-commit-id: b3979225a08bd45464e68d080ecef489f785a8bf