partly revert 301b2513aa3d [formerly d435fa477de8] [formerly 89c237804bcd [formerly 984e68f39362]] [formerly b0241676fec3] [formerly 79c3751fafd5] [formerly 8d42175cb748]
this should restore the behaviour that players don't loose if they have
no units left after the start event. Because it wasn't tested enough.

Note, that in most cases we don't need the check_victory in
playmp_controller::play_network_turn because we have a check_victory()
in the end in run_in_sycned_command


Former-commit-id: 49eaa2da824653b48745e95ce062a63197a717a2
Former-commit-id: 6215a2d375f777a1d0308a6c162584cac0fb9a5d
Former-commit-id: bebc158f0a9136fe9b097ac1c04c8ecdc5afd0d3
Former-commit-id: 803e79de852c88b184bee487534605dc0a36b07a [formerly 0b719df35132f1db68f5751afedff8372a4f5277]
Former-commit-id: e94672f64d14d3044fa4f363f348f48aa3fed567