fix 21903
it's now impossible for non hosts to advance to the next scenario before
the host does, this is the intended behavior which didn't work before
because the hotkeys system automaticly sets the endturnbutton enabled if
can_execute_command(turn_info::HOTKEY_ENDTURN) == true.

We also make the clients end linger mode immediately when they receive a
PROCESS_END_LINGER, to prevent receiving data of the next scenario while we are not playing it.


Former-commit-id: 40e8723efe60b9febe2233f0985587620ddd8641
Former-commit-id: 59ed30631a747380e758971605af5aa417a8d577
Former-commit-id: faf61123cec6c77115d6c44b8095a70195dc9d33
Former-commit-id: 4da9560eedd6eb04e7238e1653f9d781bd19f19b [formerly 61b64e24e1a9423a915b037724397e7207d2184f]
Former-commit-id: d99420ae5f78c8872b2620372cb2abf586b29a8f