fix 21903
it's now impossible for non hosts to advance to the next scenario before
the host does, this is the intended behavior which didn't work before
because the hotkeys system automaticly sets the endturnbutton enabled if
can_execute_command(turn_info::HOTKEY_ENDTURN) == true.

We also make the clients end linger mode immediately when they receive a
PROCESS_END_LINGER, to prevent receiving data of the next scenario while we are not playing it.


Former-commit-id: c569028c3af6124e19cb6e97ac0660769b52c5a5
Former-commit-id: 461e9c15e7df42c0f9158d5f115d2b3cfaeacb40
Former-commit-id: b5fba3bbd8e34ac258e6540b343f91904d28d243
Former-commit-id: f9f9d1954007fd8a0d4e9cbe06e6b2c9bd9c43e0 [formerly 59cd02b524f86a172b5d2120050d6438221c352e]
Former-commit-id: efed091da7ebf4ff443098dd10977472ddb5381d