enable msvc bool conversion check.
we now enable warning 4800. I ran though the msvc compile with this
warning and i found some (in overlay.hpp, unit.cpp, whiteboard/move.cpp)
unintended implicit bool b = cfg["attributename"] casts that use
the implicit attribute_value to int cast and then a int to bool cast.
So for attributename=true b would be false. So i believe this
is a useful warning.

There are also a lot of implicit int to bool cast from the SDL and LUA
C-libraries which use int instead of bool because C doesn't have bool.
In this case this warning is less useful, but i still think it's worth
it since a "!= 0" or "== 1" isn't cost.

Plus if someone really wants to disable this warning he can still
disable it in the msvc project settings, while before is was not
possible for someone to enable this warning the the settings.


Former-commit-id: 19e0747afe137db03ee40509ae4e4d07ca1b5633
Former-commit-id: c7d7fa7a11fa437a9a20fcf4a6928230620cf619
Former-commit-id: cf1b5144653f9a9ba4e7029f9b684249e494face
Former-commit-id: f10c537712ee05cadd74e7519614c81df872b8f6 [formerly 65b79e1ab1bbfbf81223393bc9816a7c0c428251]
Former-commit-id: 8a1f351ed2990f563bd8c02ec4d6f76d39d0b82f