more robust replay.cpp functions
we now add nonundoable commands at the current position instead of
assuming at_end(), this way nonundoable replay entrys like speak
can now also be added while we are not at end.
Usually playturn_network_adapter ensures that we are always
are at_end() and i didn't get any error with the previous solution
but the new solution is just more robust in case something goes wrong.

We cannot just do this for normal commands because it might
bring the replay in disorder.
We cannot add nonundoable commands at end if we aren't at end
because in that case the user would see the message again in
do_replay_handle in a mp game.


Former-commit-id: 9777ca8ddc3675d8cd99cf9372f48aff9d67c8e7
Former-commit-id: a7a9664d205f1d0f6550c9663bace909681f54fc
Former-commit-id: 08a03f619fb5da9059c50735f58e1ca02487468b
Former-commit-id: ffc5fa2bd29b0f108cb8a672d6a08b00239c03c2 [formerly d68e0c486f29906c12c5b60e8c18be5e53381286]
Former-commit-id: 8f3518637161f20c7595be4e6a5f9ba44182ae76