more robust replay.cpp functions
we now add nonundoable commands at the current position instead of
assuming
at_end(), this way nonundoable replay entrys like speak can now also be
added
while we are not at end.
Usually playturn_network_adapter ensures that we are always are at_end()
and i didn't get any error with the previous solution but the new
solution
is just more robust in case something goes wrong.
Note that we cannot add nonundoable commands at end if we aren’t at end
because
in that case the user would see the message again when he replays it.
Note that we cannot just do this for normal command because it might
bring the
replay in disorder.


Former-commit-id: 233a5893bdbc685e8cc7ea0da42e9156b0319a30
Former-commit-id: 1007fb088a056627317a4053bb53d00a5fab9bd0
Former-commit-id: d58cde5d547aadf8dac7c0663830914b0f5d61ce
Former-commit-id: b5d68b48e72e567e2832ef8d8edc665efb99dff1 [formerly 383af7ad9e2473a43adb1086616f29adead7e7ea]
Former-commit-id: 5c64251c37b097a384fc0c0d46c1fde7b8f0c59a