html.py: format add-on description for HTML
The description text does not get rendered very well on a webpage. One
solution might be to use pre-wrap/word-wrap in the CSS, but due to
differences between browsers, that's a can of worms (at least for me, I'm
not a web pro).

So, the not-so-elegant solution is to add <br/> to every line.

URLs are also not linked in the plain text. Although in modern browsers
you can select the text and right-click, it's still convenient to turn
them into actual links.


Former-commit-id: 9f8a07a4d82ee36886240076bd3f21bf588f1dc2
Former-commit-id: 690f15dcc7d20559d3e8af35a76c304e43759e4e
Former-commit-id: c5119c489cc9403b19a48bbd5501da3478baaf91
Former-commit-id: de6ff96e120b7c0499883af24561252029d4b18b [formerly a627ab7e0ea5dd61a2017e72fe1c342f08a85e4d]
Former-commit-id: ff2d42259c48c9f221225c83c7771cf4322d7b50