html.py: format add-on description for HTML
The description text does not get rendered very well on a webpage. One
solution might be to use pre-wrap/word-wrap in the CSS, but due to
differences between browsers, that's a can of worms (at least for me, I'm
not a web pro).

So, the not-so-elegant solution is to add <br/> to every line.

URLs are also not linked in the plain text. Although in modern browsers
you can select the text and right-click, it's still convenient to turn
them into actual links.


Former-commit-id: b7ebc4e806640bdb9afa35f8337f0961949ef44f
Former-commit-id: 2660cc59462b9876ad2f506743bcb9eb41a88022
Former-commit-id: 5e96f6adc7c4f865dfae1bad5bf501ade13b0c55
Former-commit-id: 0291c3ad187f4487b1e54070d01542791b431cd8 [formerly bc26193307c05094caa2741ec1e5170604cde8f9]
Former-commit-id: b84273ee8e88bd0b8b7243de2575b9396322755f