Insert lua_unit's contained pointer directly
Not copying the unit improves the speed of wesnoth.put_unit by two orders
of magnitude from lua's perspective.

To do this, we have to take ownership of any owned pointer from lua_unit.
We do this by not running lua_unit's destructor, but just overwriting the memory it's in.


Former-commit-id: 7aa07aa31556898db297691987cb2d6e8a3b04f1
Former-commit-id: 0801146fc803a5e147c8dd2791e77267b0fe1035
Former-commit-id: 681e1994b75361a155ff591718057fb2f99f57e8
Former-commit-id: b2325a5c6c5a651904a7d0422b70e38bb927d085 [formerly 45b9f13080c8137559b0a36632ee1eaf9f39d635]
Former-commit-id: a5ca7555e3a242f222f65bc8dc195ba53a4784ee