Insert lua_unit's contained pointer directly
Not copying the unit improves the speed of wesnoth.put_unit by two orders
of magnitude from lua's perspective.

To do this, we have to take ownership of any owned pointer from lua_unit.
We do this by not running lua_unit's destructor, but just overwriting the memory it's in.


Former-commit-id: f5f5fc1245c5789f9ba799a5911f6a53f525beb4
Former-commit-id: ff917befe3a1cb1ec9b4554bc2a566a98da4003f
Former-commit-id: 29bc010ef13a75bcfd2138c3e03a67d9bd526b6c
Former-commit-id: e33c08c35dd8a797af21376581acd9d0fe9ae64e [formerly 9fd07761fee5be0c7ef4fd8b278306542a536092]
Former-commit-id: a0f4a90a75e9055b86dc2a489e2a7117a8d23b7e