fix implementation bug in random number generator
Our simple_rng is described in comments (in .cpp file) as coming
from a desription in man pages, man rand(3). Here is a link to
that page:

http://linux.die.net/man/3/rand

Our implementation did not match the specification, because the
rand_pool_ variable was an unsigned int and not an unsigned long.

Since the rand_pool_ should basically be constantly overflowing
under normal operation of the rng, it is obvious to me that this
would affect the long term random results. It is not at all obvious
to me whether significant bias or defect would be introduced... but
it would not be surprising either.


Former-commit-id: 3f6be9d68d84a15ebffac6f6931f4b9390107ae9
Former-commit-id: e78037869b9bc805f479797598c702d02d5b8efa
Former-commit-id: 766120e6089515c66a8e0b2e0aa4ba6dea4420b5
Former-commit-id: 4a03c4df1cb2107e5373bf194a78a743b0ca4bdd [formerly 829ef54b3d2bba7274e06680c6ab789ff824fd6b]
Former-commit-id: f23903ad67e4044c53a040aeb903e4193b03e5c4