fix implementation bug in random number generator
Our simple_rng is described in comments (in .cpp file) as coming
from a desription in man pages, man rand(3). Here is a link to
that page:

http://linux.die.net/man/3/rand

Our implementation did not match the specification, because the
rand_pool_ variable was an unsigned int and not an unsigned long.

Since the rand_pool_ should basically be constantly overflowing
under normal operation of the rng, it is obvious to me that this
would affect the long term random results. It is not at all obvious
to me whether significant bias or defect would be introduced... but
it would not be surprising either.


Former-commit-id: 8410e09f6c7471d7ac06ef59dfb7cf97b22b62c3
Former-commit-id: eac6b2f49d3d80e6165863312fb08f91afeeba08
Former-commit-id: 08b279c36d5ea2292b75d7bed5ce1cf8890e728f
Former-commit-id: df35ce0b8aa6624371c0e29dfc4c18f1ec6a51bd [formerly 354708c9c38e8e894d19574333fb060f069246a1]
Former-commit-id: b1e139dfa02e43cae8990a7ed05a83eccdb644fc