style.css,html.py: use <pre> instead of <br/>; re.sub instead of re.finditer
After looking into it some more, I think I've figured out how to handle <pre>
in the CSS. So, use that, when description has more than one line.

Also, go to re.sub for turning URLs into links. The version of Python I was
testing my code on wasn't properly handling backreferences in the replacement
string when in the form "\#", causing me to use finditer instead of sub. But
I've discovered that it does handle backreferences in the form "\g<#>". So
switch to much simpler re.sub code.


Former-commit-id: 5a660f976c4db34fbdc5eaf0db213ec8a3f4b0cb
Former-commit-id: 7dfac727ffb1bfdd609aab0354c8c058f485c0ba
Former-commit-id: 7d68e1203af4e8659197561cc9dbaa43dca3ca0d
Former-commit-id: 39bf6162e5eb98f8ef9e986fa36d4d07b8d00e8a [formerly 8935f6d324c22fcb1b5a9c56a1099ac8abae75cc]
Former-commit-id: 9a83c2e50a3a7ba425d97e0cc873ee2eac963104