style.css,html.py: use <pre> instead of <br/>; re.sub instead of re.finditer
After looking into it some more, I think I've figured out how to handle <pre>
in the CSS. So, use that, when description has more than one line.

Also, go to re.sub for turning URLs into links. The version of Python I was
testing my code on wasn't properly handling backreferences in the replacement
string when in the form "\#", causing me to use finditer instead of sub. But
I've discovered that it does handle backreferences in the form "\g<#>". So
switch to much simpler re.sub code.


Former-commit-id: 6b9b569c03095fee194a257cce0a3854abea8ab6
Former-commit-id: d5084503e42f71a6cb47c13a178166ac89da51c2
Former-commit-id: aa8341757d6dfc8480127d1e41a404aa104cc0d8
Former-commit-id: 43a50429405bc06663305931c1454227b1af0348 [formerly edaf96e8ec764011ac40b5a1112744767e992b1f]
Former-commit-id: bfd0fb48b35e57fc01bfa06a694ddce2cc7fbeb7