html.py,style.css: use <pre> with no top-margin on all descriptions
After my last change dealing with this issue, I noticed that descriptions
with <pre> had an extra blank line at the top. Adding top-margin to the
CSS file made this go away, but it also made the <br/> superfluous. Thus,
it is simpler just to have every description use <pre> instead of <br/>.

This should finally resolve everything having to do with the add-on
descriptions.

If I'd noticed that the re module hadn't been imported, I probably wouldn't
have considered URL linking to be important enough to do so. Since I've
already written the code, however, I'll keep it.


Former-commit-id: 2eb8e823d1ce72aeac0b16461c1a9b003ca3b6fe
Former-commit-id: 8cb59b45a39ec63b93ebc3da1af4612c1bb09e81
Former-commit-id: 930b6a0835889777e273559d829d89bf3bb381c9
Former-commit-id: 0df10b9731ea0e3e3e8ca1e597b5f06490ac6ca2 [formerly c43ac84cdca8ceb56451fa809a07296411ab1e06]
Former-commit-id: 667799c6abc661f68a1586a7ca13a018563dc209