html.py,style.css: use <pre> with no top-margin on all descriptions
After my last change dealing with this issue, I noticed that descriptions
with <pre> had an extra blank line at the top. Adding top-margin to the
CSS file made this go away, but it also made the <br/> superfluous. Thus,
it is simpler just to have every description use <pre> instead of <br/>.

This should finally resolve everything having to do with the add-on
descriptions.

If I'd noticed that the re module hadn't been imported, I probably wouldn't
have considered URL linking to be important enough to do so. Since I've
already written the code, however, I'll keep it.


Former-commit-id: 1aebd67bf183cb7d0251644d14a5fb5b3d5224df
Former-commit-id: e9be737f75ce56ea2b50cd25061cb1a1ba384d37
Former-commit-id: 2e3b5edec47ab650b7fe4568ab7836d03fe4dabe
Former-commit-id: 9353a4801a929906940fd4d729264a30b5d391cd [formerly fa0d3bc520f08fa5f1e3057a8995ae720694412a]
Former-commit-id: ad509ca00c2e398bfc16e8c95cfa9d130046b8e9