fixup unit tests for sdl 2.0
The idea here is to make it safe to detach the worker thread. We
replace the scoped pointer for the game_controller with a shared
pointer, and pass this to the worker thread. This way the worker
won't segfault if main thread times out and decides to return.

In SDL 2.0.2, Detach thread becomes available. I haven't tested
on the new version of SDL, but my understanding from posix
is that detached threads will simply run separately, until the
time that the main thread returns, at which point they will
be terminated.

I added compiler checks so that if Detach thread is not available,
the compiler will simply disable unit test timeouts, but otherwise
keep unit tests available.


Former-commit-id: 7637a2804cd1c113b3b7c5d957d8015d401c11a7
Former-commit-id: 67e8b81d902678b569be1b19468b9ebecf28e703
Former-commit-id: aee86a2119f7ffc40b153aafb8d89442fd40299b
Former-commit-id: 79b8994a232f2f10b6162bc12c3ba8e8421a42d6 [formerly 235608745447eb3598938cdf800d866ad3c7f3e6]
Former-commit-id: ed6fbc4c6e9d81bb4950ebc2dfd62b92c6741ee6