redirect noninteractive ai test info from std::cout to log channel
I named the log channel "aitesting" and in three files,

playcontroller.cpp,
play_singlecontroller.cpp,
ca_testing_recruitment.cpp,

bound this to "LOG_AIT" and substituted where I saw std::cout.

This macro could be redefined to std::cout to restore previous
behavior.

Feel free to rename the channel if you don't like this name.
The reason I made this change is so that this info does not
appear when we are running noninteractive unit tests.


Former-commit-id: 724cff2493e122e33e72b984fe673b389a621d70
Former-commit-id: b6030185cc77c5c108e501796bb112568fc3f9a4
Former-commit-id: 0e26e2015e0402abf7598ebaa1c7dfae161b9a48
Former-commit-id: 8761dbb08c76f4e1888cc13ed7e225b83b933bde [formerly 10c6d3158363258a2f996cb5e0a2fd616ac87dc8]
Former-commit-id: cbf2d5e833114957bad8dd176ec1b040e1158d02