add wml commands 'issue_attack_command', 'issue_move_command'
This is the initial commit for these WML handlers.

The purpose of these is to provide more or less raw interfaces
to the (synchronized) implementations of attacks, moves, used
when a user clicks to make such an order, or when an ai issues
such an order.

This is useful for creating noninteractive tests, as it allows
to test the outcomes of these kinds of things without having
to figure out how to coerce an ai to do them. Sure, you can
make the same effects other ways in WML, but the point is
to test the code path that is actually used in a game.


Former-commit-id: 5baa9ae933b797be8554d8033cc589d116495f12
Former-commit-id: 9209c5f86f430b267986d15478c06638bb9b63c4
Former-commit-id: 3a886285a6e9354e2c980af104d8b8694bc2b1e1
Former-commit-id: bfdfc2403fe14655f2e4cee450e3d2eebc165315 [formerly c9b5a8d206a7e6237dfbc512d8939b3ac51b26b4]
Former-commit-id: e39c14f578585c5ae2b707201e29a60046e02a16