enable linger mode to unit test macros ASSERT, RETURN
I initially had linger mode off in the endlevel tags of these
macros, to make sure they didn't block the end of noninteractive
scenarios.

But it turns out its better to have it on. It doesn't block the
tests when they run noninteractively, and it means that if a test
fails, you can boot it up with --showgui, and then run :inspect
from linger mode to see what happened. If you don't have linger
mode, then even with --showgui the failed test closes out too
quickly.


Former-commit-id: 2c14c7ed25436922b700c69436d4e5c063ad0530
Former-commit-id: 9f9c7e34b4cfec2800ae1243e8304931be574ce5
Former-commit-id: 145210f50c187bb4997636faa2e363351256fc9c
Former-commit-id: 012e1a71ec91771876eae0c0a5ba3661c2df9e2f [formerly 845da5d42f348a75af491b1c6b88eb616c2916ea]
Former-commit-id: 5ffa320d43d29ae3a5b2e6806b5a54aeea55b68b