add unix timeout option to wml test script, enable empty_test
Adds an option to the new wml test script, to use the unix
timeout command instead of wesnoth's built-in one. This is because
the built-in one is very slightly unreliable... there's a reason
killthread was deprecated in SDL unfortunately. Will leave it for
UMC but on travis will replace with unix timeout.

The unix timeout one seems consistent enough that I will enable
the empty_test timeout check for travis.

Also, adds a help listing to the wml test script.


Former-commit-id: a22bdf069321b88407d6ddf838f5ca1c1d15668c
Former-commit-id: c597372ba6edcc9eb6c2b374860bb6de4fdc66b7
Former-commit-id: bb3a4c2f9762f236fbc173f50cfdd4c10bae48bc
Former-commit-id: 35e3e12092fed1355dfcd39de55edb5051a46ef6 [formerly 1a82dbd8df3973fc38cfcd9017d35a6e3632be3f]
Former-commit-id: f80d734287d157d76b5a9f680c224628bc6df9d6