add [do_command] wml action tag
[do_command] executes commands like attack, move, recruit, etc.,
similar to those that may be found in [command] tags in replays.
These are executed in game immediately in a synchronized context,
similarly to as if a human or ai controller had ordered them.

It is also possible to use the command to fire a WML game event,
or to execute lua code from a synchronized context.

Only some of the children of [command] are valid in this tag.

With this tag, [issue_attack_command] and [issue_move_command] are
both deprecated and will be removed shortly.

Thanks to gfgtdf for both the idea and the implementation.


Former-commit-id: 5ff1bd6d2533fb32c6c656f9bfa5f7c45bcb30cc
Former-commit-id: 431a38d0d43e69654782a0b6128fe8d0482133d4
Former-commit-id: d849c014a5f134bc0cd2b7ba53b92d65b3ebe704
Former-commit-id: 280bc19e2c24cefd224146f41d441d6b4ceb6237 [formerly b468c4d52aa1da16467d20d7ea601f6be327a804]
Former-commit-id: 78c8bf28698932f85f7c8f047d4fab162e15cfea