wmllint: warn about no side key in [*_shroud],[gold],[modify_side/ai]
Those who are updating from old UMC may otherwise be caught unawares by the
change from defaulting to side 1 when the side key is missing, to all
sides. Later we will have an option to turn off this warning.

Since the insertion of an attribute line is now commented out, it may be
questioned whether this still belongs in hack_syntax. However, at some
point the plan is to give users an option to insert the side= key.


Former-commit-id: 1403c0cf633887f25c5ca9db75fca5105696b521
Former-commit-id: 5b1b926292414cbc279fba278e866dd743733eee
Former-commit-id: ce2b23db38ffb00870b6f34af434237538e1a59f
Former-commit-id: 1ee4587a9f76e741dd481eb1c84dc86ea3cc07fa [formerly 20d1a2db1d8cf3465b7bc7b29642f6c0dfe4def4]
Former-commit-id: d3e68edd87001fbd10f3f4a4dbe55a2af39ae999