wmllint: warn about no side key in [*_shroud],[gold],[modify_side/ai]
Those who are updating from old UMC may otherwise be caught unawares by the
change from defaulting to side 1 when the side key is missing, to all
sides. Later we will have an option to turn off this warning.

Since the insertion of an attribute line is now commented out, it may be
questioned whether this still belongs in hack_syntax. However, at some
point the plan is to give users an option to insert the side= key.


Former-commit-id: 34f1d85d77ad125d5c46cf601ec939ea421e1bbe
Former-commit-id: 2faf20c4767e239f722635a66baeb99f3e9504bf
Former-commit-id: 59cd13d95fa44fa3e7f6ebdfd220eb5524d92cea
Former-commit-id: 5783b3bf02387d269619162e6e303829cc2cf7f6 [formerly 448bb15276fd396bc6af98150145966a1cb79d65]
Former-commit-id: b048cc1397c5c9d44482edd449c85a0a90eaf99c