fixup [command] [move] implementation
The move implementation used a function from map_location.cpp,
parse_location_range. This has the consequence that if the move tag
has mismatched numbers of entries for x and y, it is not reported
as an error and instead some truly bizarre, although not "undefined"
behavior ensues. Since this has obvious potential to conceal bugs
and no obvious benefit, I have changed this to a more sane
implementation. See chat log today with gfgtdf for full details.


Former-commit-id: 5c3244068c25476f0e6b9b93cb7ee30240f80879
Former-commit-id: b8cf059b5158fb128518c69a3e13a0428a265a6c
Former-commit-id: bb4f4c89f7a344d84c78630eb770f5bdfa25aa1c
Former-commit-id: 24c320524b7614b89ad11f71f642570b633f7a76 [formerly 998ccfed1fc40ec2f0d39001d2ea37c9be8bab11]
Former-commit-id: b736f0d4f226775bf26846340a13cce5c3b9e65a