fixup [command] [move] implementation
The move implementation used a function from map_location.cpp,
parse_location_range. This has the consequence that if the move tag
has mismatched numbers of entries for x and y, it is not reported
as an error and instead some truly bizarre, although not "undefined"
behavior ensues. Since this has obvious potential to conceal bugs
and no obvious benefit, I have changed this to a more sane
implementation. See chat log today with gfgtdf for full details.


Former-commit-id: 2e5205f3c88754c788efb9510d3bbdb3ea5e4768
Former-commit-id: f6588c3d3079b4b26b5401456f1ac80184ce7076
Former-commit-id: a963fc39e22f358abed5004d4466e1384a76eaa8
Former-commit-id: a02aff886b4831444b8bfb8091b326ff24a0f048 [formerly 155ac86ca015aab707bb488ace12c61b833212a7]
Former-commit-id: bb2330d8a4cb1c00f69a608652409f5573c23ed8