Fix map_locations.cpp read_locations function
read_locations will no longer silently process a mismatched list of
x and y integers as a valid path, instead it will throw a
bad_lexical_cast. gfgtdf and I discussed on irc, we don't believe
that this will break save or replay compatibility, and it could
potentially help to catch possible OOS bugs in the future. I am
currently passing all my C++ and WML unit tests with this commit.

This commit also changes the [command] [move] handler over to use
this function directly without checking for mismatched lists, so
if this commit is reverted that function will still work correctly.


Former-commit-id: 9f5af9b2e58d2f0469be58532ad4492475c7c4ca
Former-commit-id: 2f88f010a849b6604722bd2d472f10278c2d16d3
Former-commit-id: 984d15bc0483d820f487caa13375f21b8706d21b
Former-commit-id: 863ef86526a0aff9ef4a99ee76af030b7329b99f [formerly de576dcb090b048b6864b065cbec74b7d387d2b9]
Former-commit-id: 6c9549ddf9e399d79d52058ca951abbae57f6e9a