Fix map_locations.cpp read_locations function
read_locations will no longer silently process a mismatched list of
x and y integers as a valid path, instead it will throw a
bad_lexical_cast. gfgtdf and I discussed on irc, we don't believe
that this will break save or replay compatibility, and it could
potentially help to catch possible OOS bugs in the future. I am
currently passing all my C++ and WML unit tests with this commit.

This commit also changes the [command] [move] handler over to use
this function directly without checking for mismatched lists, so
if this commit is reverted that function will still work correctly.


Former-commit-id: b5b96022ff2e6502120629c3ba8623234a338dfc
Former-commit-id: b218ff18eb24219dca1b8e551d26f9125453e1d4
Former-commit-id: ac58bcbf003071e8c076c845e60b4b9eea550f84
Former-commit-id: c6830e4507ec4a6c419fb1a8af5207abf0927627 [formerly 64009cc2e4dd29fdbd5bef3224eab2a59c94777a]
Former-commit-id: 1cc27e6b9742b7ccace904eba410ba9edc13be3b