in debug mode, add team strings as well as names for hidden sides.
This brings the team names in line with the player names and other
behavior in mp_connect in debug mode. For instance there are other
lines like this:
https://github.com/wesnoth/wesnoth/blob/866847b3aa47db7205bedd6247efccebe708b9fa [formerly 716d6e1c1e9b914602beda5e31dc43250ce89976] [formerly 393f607a1796b2bed9565c28705d1b4ab92e74d6 [formerly b48d9272d877efd01d8d6a059e291be91f63c1ed]] [formerly b448552b47b007c819bf61f9105ace01ad2a495c] [formerly ece25cff97f83ddd5076a524269b736b96908424] [formerly 7574ec5100343b0a5368f6df5ba7465da788f99c]/src/multiplayer_connect_engine.cpp#L155

From inspecting the save files generated while testing this issue,
it's not clear if it could have actually caused a problem in game,
but it addresses an issue raised by SlowThinker on the bug tracker:

https://gna.org/bugs/index.php?21800

and does indeed fix a minor shortcoming of debug mode in mp connect
engine.


Former-commit-id: ca097cbcb81bda675901c8ef2ed09a441fa604a7
Former-commit-id: fbc9b9a910ea7e2e666d3ee4c28fc16298147c1f
Former-commit-id: 314e74a14653821f4047e46f1667f8bff88bb35c
Former-commit-id: e0c29cfd5640615f325dc1323583b8f235330844 [formerly d3cb8400a82557c375c81a1c7b38e4970424d253]
Former-commit-id: c23b15b8772f2337aba2c4fb21eeae247272e2fa