in debug mode, add team strings as well as names for hidden sides.
This brings the team names in line with the player names and other
behavior in mp_connect in debug mode. For instance there are other
lines like this:
https://github.com/wesnoth/wesnoth/blob/866847b3aa47db7205bedd6247efccebe708b9fa [formerly 716d6e1c1e9b914602beda5e31dc43250ce89976] [formerly 393f607a1796b2bed9565c28705d1b4ab92e74d6 [formerly b48d9272d877efd01d8d6a059e291be91f63c1ed]] [formerly b448552b47b007c819bf61f9105ace01ad2a495c] [formerly ece25cff97f83ddd5076a524269b736b96908424] [formerly 7574ec5100343b0a5368f6df5ba7465da788f99c]/src/multiplayer_connect_engine.cpp#L155

From inspecting the save files generated while testing this issue,
it's not clear if it could have actually caused a problem in game,
but it addresses an issue raised by SlowThinker on the bug tracker:

https://gna.org/bugs/index.php?21800

and does indeed fix a minor shortcoming of debug mode in mp connect
engine.


Former-commit-id: baec050bb94b0a9bc9d6a6f94e5dfa0177162392
Former-commit-id: 4016a1b999028d54d2cc88a3be959e8f2d45bc89
Former-commit-id: ea06c4abfc50ecb0b13991d53cd1c54fd7485cbd
Former-commit-id: e80a7fb0083ed476493cb36d9a7bdf6274055a97 [formerly 8b83173097f6a4786c94b85416102df1acf5d104]
Former-commit-id: 3423ec6a860f026892fdc2c43f9e15b35cc5e002