refactor is_observer out of check_victory
This is a minor change to victory_when_enemies_defeated condition.

Old logic: If there are no not defeated sides which are enemies
of eachother (i.e. there is still a battle going on),
BUT victory_when_enemies_defeated = false, and either we found
a not defeated local human player, or we are a multiplayer observer
(there are no locally controlled sides), then don't end the level.

New logic: If there are no not defeated sides which are enemies,
BUT victory when enemies defeated = false, and EITHER we found a
not defeated local human player, OR there is a not defeated remote
human player AND no local human side assigned at all,
then don't end the level.

I think this more precisely the behavior we want. is_observer
doesn't belong in this function, and with the introduction of
defeat_condition, UMC will pretty much always be able to get the
behavior they want, so I'm much more comfortable to introduce this
change now than I would be otherwise.


Former-commit-id: 1a2e7a1078b7846cdd379834c6a17f8d2f8e9800
Former-commit-id: 6daeb341225496f0fa5792da70e4466a0340ba05
Former-commit-id: a80d470ba6f80b6c5f29cddb90c64e711a6a86b3
Former-commit-id: c6dd715d412a630f7b3f1e3e9b37426c044b8379 [formerly 9d95ec2b37cb787cf5ec3a834aa79f93047701f7]
Former-commit-id: 51ae67a1d6fc03da457c48121bf1ac031a32ba18