refactor is_observer out of check_victory
This is a minor change to victory_when_enemies_defeated condition.

Old logic: If there are no not defeated sides which are enemies
of eachother (i.e. there is still a battle going on),
BUT victory_when_enemies_defeated = false, and either we found
a not defeated local human player, or we are a multiplayer observer
(there are no locally controlled sides), then don't end the level.

New logic: If there are no not defeated sides which are enemies,
BUT victory when enemies defeated = false, and EITHER we found a
not defeated local human player, OR there is a not defeated remote
human player AND no local human side assigned at all,
then don't end the level.

I think this more precisely the behavior we want. is_observer
doesn't belong in this function, and with the introduction of
defeat_condition, UMC will pretty much always be able to get the
behavior they want, so I'm much more comfortable to introduce this
change now than I would be otherwise.


Former-commit-id: 45e0210df7b99d2a58d87dcc23666d8df7e55443
Former-commit-id: 1491318ba08961165415cf2ad0494f42b3260932
Former-commit-id: 78720301c8e4fe828652bd82ea2726f5860e2631
Former-commit-id: 87c925658bd394af9b89be5342e61e09e09042be [formerly 64b785a1f3c7b761f3bd3946c41b059b6f9ef73b]
Former-commit-id: 69cf4c2359c4be6b643f661b27804efc06940f9b