Move animationless moving unit hiding to a more appropriate place
Unit hiding is usually done in replace_temporary, which is called by
proceed_to. As proceed_to exits at the beginning if animate is false, this
would seem like a good place to hide the unit, but it is also unhidden during
this function. This means that was_hidden_ is set multiple times, with the
second time overwriting it with the temporary state of being hidden.
As animate remains false for the entire lifetime of the unit mover, we hide it
at the start and let finish unhide it again.

Reported at http://r.wesnoth.org/p570866


Former-commit-id: c85807751030618632810f5f9dc53771183658c4
Former-commit-id: 404a75b51c0310183cba7ed3c3a6a796693f6712
Former-commit-id: bf636a920ed6379249daec94fa42492baf57b8c6
Former-commit-id: 45b58b61c9de9e65f5b5ba41a92d2dec9af5d8ed [formerly d498abb7a059c2090681d1760d1556463d2989fc]
Former-commit-id: ecb6ddf2c400aff4446dccf246d41d5724ba4949