Move animationless moving unit hiding to a more appropriate place
Unit hiding is usually done in replace_temporary, which is called by
proceed_to. As proceed_to exits at the beginning if animate is false, this
would seem like a good place to hide the unit, but it is also unhidden during
this function. This means that was_hidden_ is set multiple times, with the
second time overwriting it with the temporary state of being hidden.
As animate remains false for the entire lifetime of the unit mover, we hide it
at the start and let finish unhide it again.

Reported at http://r.wesnoth.org/p570866


Former-commit-id: 5c6e2e3d25061a9f26d786481968086a2862276e
Former-commit-id: 4254a2ff20b31634e29693a6dd5dbaddf27545ff
Former-commit-id: 15c49913437dffb33b802774b043c1774c354eff
Former-commit-id: abb5d01075b33adcd22bde0ab9f97ac043771580 [formerly 13aa3af314e4bf964fcead02c26489e1c9c8825e]
Former-commit-id: 37de7957174ce37c62daa6326f985362c557aef1