add error reporting when [if] is evaluated with no actions
I frequently make the mistake of using {IF_VAR} but forgetting to
use [then] with it. One solution is to use my own macro, but the
better solution is to add proper error reporting to if. If the
engine parses an [if] tag with no [then], [else], or [elseif], it
should flag an error... otherwise these mistakes are generally
silent and very difficult to find.


Former-commit-id: 3d4440412c691567dd282bdedeb80b2c5ace2fad
Former-commit-id: 5d3f61cbf9ecbb08a4aefb4c492f3316ba6bef8d
Former-commit-id: 0319b608ed5c5ad897130c46d537763b95373c89
Former-commit-id: 3d561bf106138568e6c74f374738eac8a128fca2 [formerly a98e06b2fa90be7f360b1b2753d3b6bc151a21d5]
Former-commit-id: 2f0f1ab343c108b10af5354c955d172662ee550b