Windows WML tests: Add WindowsTimeout project
As the --timeout option for Wesnoth can stall the process, we replace it
by an encapsulation comparable to the Unix timeout function.

WindowsTimeout isn't limited to Wesnoth, it can in fact set a timeout for
any executable called through it.

An example on the use of these WinAPI functions can be found on
http://msdn.microsoft.com/en-us/library/windows/desktop/ms682512.aspx

Former-commit-id: aa97eb72b75c2f1dd361474bc176d60ad4b941ee
Former-commit-id: 8985e4392cf4a5ea20a97ed16dfdcd87095441ea
Former-commit-id: d7f979ed395534a17257d3afbe0aa942fb826082
Former-commit-id: 1ddd8459db0bf23fddd54b51f7b1b41b965bcbba [formerly b5c6f689618c18bde4b2d8a87ab9e37beee8ff6d]
Former-commit-id: 1c0590b3df9ac44fec205b2b7c59c70286b41499