remove built-in wesnoth timeout
This was buggy and based on deprecated SDL features that will not
continue to be available. aquileia wrote us a very nice windows
port of the run_wml_tests bash script, so for windows mac and
linux we should now be able to run the unit tests as desired.

Maybe someday built-in timeouts can be supported using boost
process library, if that ever materializes.


Former-commit-id: d45cce2ce7d619ee7ad08d4a5164e8ffcba26cba
Former-commit-id: a9a6496f784f57661889146d6c2526bfa27d54c1
Former-commit-id: d8583f39e2a25279404045810da091e4a6ef5350
Former-commit-id: b4da7dc0328dccf8cd411bdb805414b0131c4514 [formerly 83b1b5259934a479a93fdb9621c0821822c070a4]
Former-commit-id: 11bcdc69500e33e52e01c62e46de7f561c1543e1