remove unnecessary arguments
All of before_human_turn, play_side, play_turn had an argument
"bool save". The only use of this was by the playsingle controller
to determine whether or not to save an autosave -- it was controlled
in a single place by a single for loop in play_scenario.

Instead of polluting all of these functions, including play_controller,
and replay_controller, with arguments that they ignore, save was
replaced with a member variable of playsingle_controller. This is a
strict refactor in terms of behavior.

Conflicts:
	src/playmp_controller.cpp

Conflict was from a replacement of "init_turn_data" with
send_turn_data which is not being backported to 1.12


Former-commit-id: e92aeb266cf864a74b017a6f27ac476c46134b0f
Former-commit-id: c7e0ad36c190faf8a5a3b23a26ad3f395517dd4f
Former-commit-id: eefb367587ad7d02959e23225c2a773baf2e68d9
Former-commit-id: 4a28b4f42a0374718f4894c1ca88dc2162abd68e [formerly 5c664f0bf9855333fe1999b0a2820d614e07f73c]
Former-commit-id: c71443e56ac2ec23a3d03ba22e7b2237969ce5ec