remove unnecessary arguments
All of before_human_turn, play_side, play_turn had an argument
"bool save". The only use of this was by the playsingle controller
to determine whether or not to save an autosave -- it was controlled
in a single place by a single for loop in play_scenario.

Instead of polluting all of these functions, including play_controller,
and replay_controller, with arguments that they ignore, save was
replaced with a member variable of playsingle_controller. This is a
strict refactor in terms of behavior.


Former-commit-id: cea87108270752754031621af21c106001431e46
Former-commit-id: 0d4702d2890f70bc8bba9f356e9d987411d5b6c0
Former-commit-id: fd0c6a88e72461750d2c4e3d1fd9419ff40ec0b2
Former-commit-id: 64d1748ec408bfced76468c548ff4a442d6c5417 [formerly a4d95a39d3894c80d655a3813e289e7c2c455bb8]
Former-commit-id: d4cf22213e05f296a263802d3d62418aa4bb5485