remove unnecessary variable "turn_over_" from play_controllers.
After other refactors, this variable is basically useless -- it
is only used in one function, and indicates to that function if
we called it from the normal sp path, or the mp linger code path.
We choose to replace it with the linger_ variable -- the sp code
path will never follow to this function if linger_ is true, and
if it did it would cause a serious error anyways and corrupt
replays, so we would know if this was happening. So this is a
strict refactor.


Former-commit-id: 1eb315aad2d737f671280d6614ebe59e7e0341ab
Former-commit-id: d1755070f9eeb118512f934de620b967e2823295
Former-commit-id: 2849524e75a879dbf7291b52c5399b920fa5105b
Former-commit-id: 74fac1b6436040d6eab9ff25db1374d3afa22da4 [formerly 74f66f71b5b31de07f2759e1ff8986b8a82b7a6e]
Former-commit-id: 3256f73cc7142fa40d3ff2fc6da935a2e471dfc3