use completion setter object in mp linger mode
mp linger mode was not properly setting the completion status,
as sp does. This caused mp end of scenario saves not to end up
in linger mode when they are reloaded.

Note that this reopens bug 15545 (!)
https://gna.org/bugs/index.php?15545

In 1.11+, I believe that the correct way to fix that is to save
the [carryover] while in linger mode, and restore it on reload,
however this is for another commit...


Former-commit-id: d541e78ff62af206e3ba394e7d818b9322cfff64
Former-commit-id: 809c54f6d68046419a0b2bdcf63f7660c44db45c
Former-commit-id: b0a16137173157193636709b9d854bebe0e57311
Former-commit-id: 7f920725ad2b7a1b659a417c765ad70aefc9f8f2 [formerly 7bc82e056ce83706e172e39e7075ac42f37c39ac]
Former-commit-id: 32277f6b580313d68981fd5909f66ed6db787dd6