replay controller uses game_board to do resets
This is not a strict refactor -- the game map will also be reset
to the starting configuration now, when we reset the replay. It
doesn't look like this was the case in the old version, so it's
possible that if a scenario used [replace_map] or similar, the
replay could go out of sync after hitting the reset button.

Regardless of whether the old version was buggy, the new version
should be just as good if not better, and was tested.


Former-commit-id: d5125bbfb6dda352615debd37c3a6291393a7287
Former-commit-id: cbeda04bfc76b9b33ac88efc1e73647b976110cb
Former-commit-id: 72b63faf5d98de89b6318a9c00e32e32397766b9
Former-commit-id: 3b896e39b755e3539c9e7be8acef926130fc2095 [formerly 6161ff090a98b0564bb8e3cb498eaa398a861e9e]
Former-commit-id: 18ace03e2b2b1b0081c96559edb686b2d32aa103