remove "parent" from classification
"parent" is a leftover from a feature that was planned 5 years ago but
was never implemented. To have something like a "tree view" for
savegames in the load dialog.

"parent" gets written but never gets read.

The current code writing the "parent" doesn't really work because it is
only takes the name of the last savefile unrelated of whether there was
a reload, or similar between so for example a "parent" of an sp campaign
save could easily be a mp match save. Also it could produce cycle which
would also be wrong for a tree view.


Former-commit-id: 427c88337633c5409748c10b405121a21f346704
Former-commit-id: 8e5c46c6d114c4c63db06ac03fc0e0f73bcc56bc
Former-commit-id: 46edd803994053c58f824afbfb4813d30959fa18
Former-commit-id: fcfd5a758b731c8433bba9d023c5f9a460b6da31 [formerly ba79f5295504f69618a069d605f40676ed2eb969]
Former-commit-id: 1f92e3b157816695ca7a50e66f8101c932f8ff95