remove resources::state_of_game
we don't want to expose the fileds "snapshot", "carryover_sides".
instead we only add classification, and mp_game_setting to resources
which are members of game_state.

in order to do that, we have to add a do_autosave method to
play_controller.

this commit also moves some code from game_state::write_snapshot to
play_controller::to_config which was intended to be in a different
commit.


Former-commit-id: 42650a7c5d454eb90e828073ea7a827533557fed
Former-commit-id: 1a6724b1aa0a13ad98e05e8e34032c85d15fefb7
Former-commit-id: 291f05e1232ef4e842aeb92f8da29a2e3d1cc35a
Former-commit-id: d9ace7e7aa6c3db7d13281554dcdc51d797532a1 [formerly db0095300d3dce526aea6d45b8b1e12593559aab]
Former-commit-id: 85599bd583d64bde60ee5a45242e859367e8ca5a