integrate game_board into turn_info object
This commit manages to purge all references to resources::teams
and resources::units from the turn_info object, by moving
functionality to game_board, or refactoring to use const &
accessors.

This is not a strict refactor, but I had my hand in the earlier
state of this code anyways, and I have playtested the new version.


Former-commit-id: 508a6c8b8f1b6b024b4dc41cdfb9571d0970a7c8
Former-commit-id: 6972b6e0d6107c24274f96511b93f7d29a7f2f88
Former-commit-id: 3c46ee311fddcff4cad8be28cdb5cb172b66e82f
Former-commit-id: ee16b91386414f183e61b82b2edae9ffc958b498 [formerly 962fdf7df588dd38387f20aed15b00b0fac4b076]
Former-commit-id: 52bf4f9948fff08d5595807801bcdf53a884af0b