resources::game_map is const *, move functionality to game_board
This commit changes the resources::game_map pointer to a const *,
so that changes to the game_map coming from lua or wml apis must
go through the game_board. Three functions were moved into the
game_board -- although substantively they did not change, there
were minor changes, in terms of the kind of error reporting done,
and also in how gui refresh orders are issued afterwards. In
testing the differences due to this refactor don't seem to be
noticeable.


Former-commit-id: 936ac3c89d027e289ff379079601c7f0cc4e5aa9
Former-commit-id: aaa2a4728229b9110b024ea086f01235615e39bd
Former-commit-id: 52ca421ccccf28b3b83c1040c87c0aad478a631c
Former-commit-id: 953a71b0501871f80a13b96dba094246f50f6dd4 [formerly 71ff97eb4e4043075b5c029e1e757309b1ddba36]
Former-commit-id: 985ced8f07d8213e8865b060ad41c8a6d0792761