campaignd: Quick-and-dirty implementation of upload-time blacklisting
A path to a blacklist WML file may be provided in the campaignd
configuration file (server.cfg). The BL file, in turn, may contain
attributes for comma-delimited lists of wildcard patterns, accepting
shell-style '*' and '?', by leveraging utils::wildcard_string_match()'s
functionality.

The following lists of patterns are recognized at this time:

  ip=<numeric IPv4 address masks>
  email=<add-on author email masks>
  name=<add-on id/name masks>
  title=<add-on title masks>
  author=<add-on author masks>
  description=<add-on description masks>

Currently, the IP address mask list also takes '*' and '?' wildcards. My
intention is to use CIDR subnet masks instead of or in addition to this.

The blacklist WML file is only read by campaignd once at startup and
it's never written to, thus allowing to preserve any WML comment lines
that may be included in the file by admins for convenience.

The IP, email, name, title, author, and description of a (new or
existing) uploaded add-on are matched against the blacklist in that
order. If the add-on matches, the upload is aborted and the user is
shown a generic message in English, just like with every other possible
campaignd-side error (on the plus side, this means it can be backported
to 1.12).

All matches (except for IP matches, which only contain digits and
punctuation) are done case-insensitively.

Because of campaignd protocol limitations, the upload is only checked
and aborted after the client has already uploaded a weighty WML document
including the add-on archive. Such is life, I guess.


Former-commit-id: 1a35a78ed3fe8fead093d824ad19f92456e1cba4
Former-commit-id: 4f14cfd1119061e38c7f09b8b4dc04c1c12e79af
Former-commit-id: 32b769e9f6931524b6eaa6e2f135534ad6358ed6
Former-commit-id: e3c2c32b670061e2f164254d27f5b358ca57ebb8 [formerly bd375f46bc6adfc486e741f1f1107531c93cd088]
Former-commit-id: 952378a1e5ffb87dc837e6c8b1b0a4374fca2fa8