campaignd: Quick-and-dirty implementation of upload-time blacklisting
A path to a blacklist WML file may be provided in the campaignd
configuration file (server.cfg). The BL file, in turn, may contain
attributes for comma-delimited lists of wildcard patterns, accepting
shell-style '*' and '?', by leveraging utils::wildcard_string_match()'s
functionality.

The following lists of patterns are recognized at this time:

  ip=<numeric IPv4 address masks>
  email=<add-on author email masks>
  name=<add-on id/name masks>
  title=<add-on title masks>
  author=<add-on author masks>
  description=<add-on description masks>

Currently, the IP address mask list also takes '*' and '?' wildcards. My
intention is to use CIDR subnet masks instead of or in addition to this.

The blacklist WML file is only read by campaignd once at startup and
it's never written to, thus allowing to preserve any WML comment lines
that may be included in the file by admins for convenience.

The IP, email, name, title, author, and description of a (new or
existing) uploaded add-on are matched against the blacklist in that
order. If the add-on matches, the upload is aborted and the user is
shown a generic message in English, just like with every other possible
campaignd-side error (on the plus side, this means it can be backported
to 1.12).

All matches (except for IP matches, which only contain digits and
punctuation) are done case-insensitively.

Because of campaignd protocol limitations, the upload is only checked
and aborted after the client has already uploaded a weighty WML document
including the add-on archive. Such is life, I guess.


Former-commit-id: cbc23ceadf5a439d7bd3d8f0762ee62ec87fdd45
Former-commit-id: eea8bddd34f399244c3a4fd22647a7da91760627
Former-commit-id: 075c48d0c4adfdc044a5864804231d9770e3ee4d
Former-commit-id: 81818dada26b6633007acb4c8c7c711376f8ebe3 [formerly b8d94e5b9041229fe19db74cd37efd2f85942e90]
Former-commit-id: 24467ce52013fe7b5265e38064307afce446f21f