add end_play_signal type and visitors
This commit adds an "end_play_signal" type. It is meant to be used
in play_controller classes to prevent crashes due to bad exception
handling / collisions, and should help us to make assumptions more
easily about when networking operations will happen

It also adds some macros to avoid the proliferation of unreadable
try catch blocks. It is expected that many of these can be
expanded and simplified after the refactor.


Former-commit-id: beea16aeb9d9cae382df5952e4fbae54a9e5b138
Former-commit-id: 9c989d0f859978f2286463057e1f63f22b8bae5b
Former-commit-id: 0107ba1e2860056402d9959550fea400b6ef5823
Former-commit-id: cc977131a9b7e3c749b3541892ba6bca7b3f1963 [formerly b7c9cd664b5d2ff9bcbf294681305d797ce280b3]
Former-commit-id: c840e78793adc77941e7c5e68c546acf5737ba2c