push end_play_exceptions one down in call stack, at "play_turn()"
Continues trend of previous commit. play_turn becomes no throw
with respect to these exceptions, and so we can switch calls to it
to use the "PROPOGATE_END_PLAY_SIGNAL" instead of "HANDLE_...".

It looks like by progressively doing this and unit testing at each
step, we will be able to successfully convert all of the functions
in play_controller to use PROPOGATE, which is a cheap if else and
not an exception handler, and increase the stability of the engine
by avoiding the use of exceptions for control flow. If we continue
this way, we could push all the throws into the event handler
and the replay module. Perhaps could continue from there, but
likely at some point the exceptions might be the simpler solution
esp. for breaking recursion in WML / lua or things like this.


Former-commit-id: de65a5a1a62b660c78649d527aad84750e04f9ed
Former-commit-id: d34f6cf0397bc76b09249dafa9cbf4b2aad4af65
Former-commit-id: 89225c0a91c80c53b5a01dfe6b54e1e98a434f11
Former-commit-id: da5ec42b4f6edb7b8f890bdb66515187c6066fe9 [formerly 6d5046ecb15a8a1f9925ac1a65b587ffff137aa6]
Former-commit-id: 2d11986718f3afc01439f7f3174f8ade7d82608f