gui2/tgamestate_inspector: Use write() function for printing WML objects
The write() function from the WML parser is a dedicated WML
serialization mechanism that is also used to write saved games and such,
and thus it will output WML in the same format that the game would
normally use. The config::debug() method we were using here before,
instead, only implements a simplified strategy that will not produce
valid multiline attribute values, for example, or include any textdomain
directives applicable to the contents.

If we are to have a way to copy the inspect window's contents, it should
prove far more convenient for coders to see the generated WML in the
same format as it would normally be saved to disk.


Former-commit-id: 748b1167bdad90e9f3dbd0d1402fb74438d94540
Former-commit-id: 2f6c2bf68ceed4f7c6043b89a1206aacf08036b5
Former-commit-id: aebc3b2002a690e7ae1393d09b29241dbe3c89f8
Former-commit-id: 2001d2cc2d591aca4631154884c2ad839a80148c [formerly 21c15f6411a4660670c3de85496f3d2755d24743]
Former-commit-id: 11f33b3233a26db7eaf59a53f8c149921ecdd66e