gui2/tgamestate_inspector: Use write() function for printing WML objects
The write() function from the WML parser is a dedicated WML
serialization mechanism that is also used to write saved games and such,
and thus it will output WML in the same format that the game would
normally use. The config::debug() method we were using here before,
instead, only implements a simplified strategy that will not produce
valid multiline attribute values, for example, or include any textdomain
directives applicable to the contents.

If we are to have a way to copy the inspect window's contents, it should
prove far more convenient for coders to see the generated WML in the
same format as it would normally be saved to disk.


Former-commit-id: a4b2a7fb06385e869644d6476b97109da150e40c
Former-commit-id: 5796f29db8f68e02d54031c09e67cb145c3ca825
Former-commit-id: e584a6ffa1af4997f415d3ce7892e992d0fc7011
Former-commit-id: a69a7248c5d50e64aa51f1ceb0df3b20b5729964 [formerly b27c6ba2d99aa2b71445454adeda8c2a193610ce]
Former-commit-id: a9fb8738d07aeb58aa1c600c4f1be8d61d8cd818