gui2/tgamestate_inspector: Add a button to copy contents to clipboard
This copies the contents of the main display area to clipboard. The
button currently lacks a tooltip because the tooltip causes map labels
to glitch through the dialog when displayed (noticeable in e.g. the test
scenario). I'll file a bug for this issue later using a (commented-out)
line introduced by this commit as a test case.


Former-commit-id: 185c816b7feda9cd6d164049ba04569ffc3f73d9
Former-commit-id: f9414beae775d3d9cee940ec184db1940997c209
Former-commit-id: b7ce89a2ddd9ed4067409f45869082fa9c928251
Former-commit-id: df600d165ef56976f570cc66683139a5e0e98c66 [formerly da091deac2f2cb48988b837ccbcbddb84ed11455]
Former-commit-id: d96fb87a8fe611dfe6e693a7c7e505e9fd59ed78