gui2/tgamestate_inspector: Add a button to copy contents to clipboard
This copies the contents of the main display area to clipboard. The
button currently lacks a tooltip because the tooltip causes map labels
to glitch through the dialog when displayed (noticeable in e.g. the test
scenario). I'll file a bug for this issue later using a (commented-out)
line introduced by this commit as a test case.


Former-commit-id: 33a6a184b2a64476dffab65a57b45b11c176e4e7
Former-commit-id: dc04c2a3969c82df85934a4c1f49b929cd13f739
Former-commit-id: c9152b3d9cbc8357f82bd63dd8cb9c46faab5663
Former-commit-id: 8fdc766c822fe6e41e6ed30dad0535da3b649340 [formerly eb21574002afdc85896b6d0bcbb6e5d4ee12f19f]
Former-commit-id: 90ad0f2400bfc780fb6700c8439fb9db44f2661a