main catches and handles "std::exception" and "...", returning 1
The purpose of this is so that even when we request the program
to be terminated with "throw 42", the exception will still
propogate back to main and all destructors will be called along
the way. This way, if the program requested a socket for example,
the socket will still be released when we terminate unexpectedly,
so it might be slightly more convenient for testing.


Former-commit-id: a388a58b24b9d9fb53b2e308babc263fd9ee840e
Former-commit-id: 6acb26ac20d4df1c8fe4d783cfbc17359bb3c5f4
Former-commit-id: 53c15f568a586726bee6f865a3f7f65d5468821f
Former-commit-id: 8e96533c4fdd261047ffbe6ff460350ca0ddf105 [formerly 53e493085816e04ebfc74d56063f480190071922]
Former-commit-id: 438ae82bd698c9a47a3f5d1ed2c830faa3b985d1