campaignd: Rename and refactor campaign_server class into campaignd::server
This is the first refactoring step.

This moves the class declaration to a separate header file and switches
around some implementation details. Included in this commit are a few
minor documentation changes, a lot of whitespace and placement changes,
and the following API and behavior changes:

 * Reordered server class fields, keeping net_manager_ and
   server_manager_ as the last to initialize.
 * The campaignd hooks configuration (which hasn't been used in ages) is
   now done in the load_config() method rather than directly in the
   constructor.
 * Both read-only mode configuration and control socket set-up are now
   done in load_config() rather than the run() method.

Included as well are a few Doxygen documentation changes of varying
scope. Most notably, the description for the fire() method's return
value is gone since the method was made to not return a value some time
after its inception.

I also took the liberty to spatially-reorganize some of the involved
code except for the fire() method's body (which is massive and had to be
reindented anyway), in order to make the code more fluid and clean to
read.

Overall, there are deliberately no behavior changes in this step other
than the aforementioned minor bits, in order to make the next steps
easier to plan, perform, and debug.


Former-commit-id: 135d13025daed310f7184bbf2b7b535bd998ff6d
Former-commit-id: 82bf89aae09584818ba61452b63651750251ee4f
Former-commit-id: a96ccddcf42350766d36bc178c80152707e16c60
Former-commit-id: cbbe6041512b8d543d5c2d48d3341a99fdf8b8f6 [formerly 01d945b897779472cd0434f0ad50da39241b9c8a]
Former-commit-id: 46bc8441a6d7cee6dcd267d2cac35fcd51c7a0bb