campaignd: Rename a server class field and some scoped_ostream objects
The file_ field is now cfg_file_ to avoid ambiguity ("what file and for
what?"). Since there are a lot of scoped_ostream instantiations around
named cfgfile which are used to write to the cfg_file_ path, those are
now renamed to simply out, also to avoid ambiguity -- just an
intermediate step before refactoring those into a separate method.


Former-commit-id: be6c53cc318d2f5843111f74ba48213afb73671c
Former-commit-id: 862bc49abaab8c5191552180d2b74748adb08ca4
Former-commit-id: a70f8f685327b3fce221b367da5d81620670b603
Former-commit-id: 2b7fecb55dbea50772bd57068a13f3c7368c51eb [formerly 8465010a0794242baf2636f2eb3cba942ee9ee4a]
Former-commit-id: 92b8c9b9ea5e4bf5e1705fe676273378a51eec26