campaignd: Refactor request handling blocks into separate methods
These methods are tried and run in sequence by looking them up from a
request handler registry maintained as part of the server object. Since
every campaignd request is a single WML node with an identifying name, I
feel this approach makes things more readable than the previous massive
deeply-nested if-else-if chain approach, but that might be just me.


Former-commit-id: 845e39d3f6b3b242780663aefe1212800fa7f151
Former-commit-id: 1bc17f2b8b06816d974bedcc5d667862dca71544
Former-commit-id: 31396f64dca9c3682f9b479ce3872d21f9d855b2
Former-commit-id: 510b259a9789a3619c883fc124ce9188d1a69396 [formerly 6740d441a4666cd5f7122d75e8015c7ec0cacc35]
Former-commit-id: 7550035daced15bfa9f9b2d78260026d15cdc6c5