campaignd: Refactor construct_message/error() functions and boilerplate
This replaces them with two methods that are part of the server class
itself.

These functions were always used in conjunction with
network::send_data() to send the generated WML object to a network
client. Because the config was created within the functions and not on
the callsite, this resulted in a call to the config copy constructor
that could be avoided by grouping the network::send_data() call together
with the WML generation.

Former-commit-id: 696cd06d86796371aac8c99d2d3944116988d6c6
Former-commit-id: 46f18cc2d8e641a080879c7c6dced0bb816c7096
Former-commit-id: 898f884991f80ed8137c0837552a294b0dd727c7
Former-commit-id: 04db9feedbd6416722b196c15a87ef57fe111a83 [formerly d53b12a89e7c9513b4b50d314efb5dc015442faf]
Former-commit-id: 5ce2a2966aebdbc60b0aad70e3b1e02940b2f179