Fix an assertion failure in the GUI2 placement.
If a vertical scrollbar is added to a widget its size is recalculated.
Unfortunalty the determination of the minimum width didn't take the
horizontal scrollbar in account. This lead to the container to be too
small for a widget with horizontal and vertical scrollbar causing the
placement to fail.

Fixes bug #22095, possibly #15615 and #21785.


Former-commit-id: 9583e5e1766e9f52ff8b6c405c0ba39d92cb1a46
Former-commit-id: 3a851b4db622d5b1c968897eed364ad6a5c362ca
Former-commit-id: 1002c94464651c9c46944d124422dd010b934f85
Former-commit-id: 48cb997dc45440bd9e8616be4b89779a6879f05d [formerly fae410d377c95fb5e0aa0b4d4d9ff577b9439707]
Former-commit-id: 348bc9d25a6c84927f5fa0a68685205a45e96437