Don't catch config::error in config_cache::read_configs()
Reverts a bit of commit 8df162913714b7153dbbc7a5368c4977bf16936d [formerly c4ec1773ab4e630ec2f3eb10284d3f278ace0ea1] [formerly 109778b443a9c25366147d1f6204c48692938c2e [formerly 029bc1460a4198892e8eb7feff21ec6edc4df8dc]] [formerly 075f251d34d9c9d036e1549c3542ab08be1a9270] [formerly 9e53a4aa3966faf970b044f5a60623587d57faae] [formerly 9184e93920f764012c11194bc059045508b3479d] to fix
a regression affecting both master and 1.12 since after version 1.11.14
which was causing WML validity errors to not abort the WML load process
or result in a message to the user in the GUI.

The WML parser and preprocessor throw config::error and
preproc_config::error to signal callers of issues with user input (WML
documents, in this case).

We need any exceptions thrown by the WML parser or preprocessor to
propagate upstream so they can be caught by the code responsible for
initiating the sequence so it can best decide how to present them to the
user. For example, the game_config_manager class will format and pass
them to gui2::twml_error so that the user can be informed of errors in
an add-on or mainline through the GUI.

Because config_cache::read_configs() was catching the exception and not
propagating it upstream due to a change that's part of the
aforementioned commit, this sequence was broken, and while the error was
reported to stderr, the game continued as normal in most cases with
a malformed half-parsed WML tree, resulting in all kinds of misbehavior.

This issue was originally reported in the forums by SkyOne:
<http://r.wesnoth.org/p571380>


Former-commit-id: 10de3521cf8af05e48033ef0994961e77871647d
Former-commit-id: cbfbd54efbda888663266fad1bf7b45d1b075391
Former-commit-id: 7f43f766d559a5544d671978450259166be2c3d9
Former-commit-id: 0059e7ef25518712c9b803e766e08ef8e45d9b17 [formerly 32655ff66956814fd68f336bbe82b541e13080c4]
Former-commit-id: 179f7512e57f4e893f745c90ae47780f2abda06f