Fix an assertion failure in the GUI2 placement.
If a vertical scrollbar is added to a widget its size is recalculated.
Unfortunalty the determination of the minimum width didn't take the
horizontal scrollbar in account. This lead to the container to be too
small for a widget with horizontal and vertical scrollbar causing the
placement to fail.

Fixes bug #22095, possibly #15615 and #21785.

There are still a few issues remaining, but it seems they only cause the
code to behave less efficient. The fix for this issue seems to be more
intrusive and will only be committed to master.


Former-commit-id: 7d303bd2d9a1e015f23fb65e914e590cfd06aafa
Former-commit-id: d042a1ddcda7941404cca4b522ca0237d9ab3534
Former-commit-id: 240555f5b11569b3191ccaf17a8f7b3446df363f
Former-commit-id: 25531e3065f8b52a06d83ca313777525c42bf6e7 [formerly a2ed7f6bcfed10acfe4e68c362269d8ea2aab940]
Former-commit-id: 8a921f4233353f65247fd5e994296d4aba096272