display object uses a const unit_map * internally
This was made possible because const units may now be drawn and
animated. There are some clients of the display object that still
call display::get_units, which is a non const getter, and which
we support by a const cast. The next step of refactor is to get
rid of that, but in this step we have at least gotten rid of all
internal uses of the non-const unit map.


Former-commit-id: cf6ae4401b8f872024833753f776681ac9a6a854
Former-commit-id: 4d820c5d3db6127824e99597888d1e13bf6aa65e
Former-commit-id: 1949e13d0507487d6697d17458d4cd85c6817d91
Former-commit-id: 59afdfe3ce70c4440de64372c02b53bf910ce7d7 [formerly 3cc466eccd736feba0e8bf29c3787fd82907ed6a]
Former-commit-id: 066188c99ff8cec2c2c0947ecde1d41199ffe764