only const unit_map in display, unit_attack fcn uses game_board
This commit makes all the necessary changes to allow
display::get_units() to return a const unit_map rather than non-
const, and to eliminate display::get_const_units(). This allows
us to eliminate a const cast introduced in prior commit.

In leader_scroll_dialog, unit_animation, and some places in
unit_display, we change the call display::get_const_units to
get_units, and mark the return type as const.

In unit_display::unit_attack, we can't do this because that fcn
also calls unit::take_hit which is part of the game logic and
should not become const. We make that function take a reference
to game_board instead and pass resources::game_board from the
caller. Also we pass display in as an argument from the caller,
to improve encapsulation marginally.


Former-commit-id: 27d0940449dacf91b65e6ef382c3f0537dd1d720
Former-commit-id: dc70c877f60bd249651c881beeb5f5830281dc9b
Former-commit-id: fd0a6e6d23e741258175e5dd570e666940e76c86
Former-commit-id: 362e43170e8147222659d1ac4f7925ac473dd03e [formerly ba05ea4b5fa90aa23f209aea582370d806a62850]
Former-commit-id: 1baee44e80f854483b2de7594eec4d7e6f3e6cbf