display uses display_context internally, not 3 private pointers
Also refactor editor and game_display to use this.

To achieve this it turns out we also have to add a "dummy display
context" unique to the editor code, which it can use to initalize
editor display, because after refactor NULL doesn't cut it
anymore. This appears in src/editor/editor_display.?pp, might
want to branch into its own file later.


Former-commit-id: 81bb98e328066039c28484de04f3cf4602e8af13
Former-commit-id: 1c66e654b3e559275ffd86c443222a71cab420ec
Former-commit-id: 1bd6264ace9e6d287d753b3381f13325fa3adc61
Former-commit-id: 4a7007c2f2d354cbdb08aa228323d9908b2720cb [formerly 6eadbcdeaf9313aee8bf28d070df07314f394d57]
Former-commit-id: 29acdfba1d9499c44fe43baa9de9efa0a6ec7895