remove resources::game_map (part 2)
This the result of executing, in folder src/, the following

find . -type f -exec sed -i 's/\*resources\:\:game_map/resources\:\:gameboard->map()/g' '{}' \;

and carefully inspecting the result.

We also had to add game_board.hpp includes in various places,
and change the arguent order of unit::is_visible_to_team --
this function was taking *resources::game_map as its default
argument, and we do not want to include game_board in unit.hpp,
as it creates cyclic dependencies. This was resolved by
eliminating this as a default value -- this is an improvement,
since usually when this function was called it was in a context
where a game_map was available already locally anyways.


Former-commit-id: 787daee8ebeecc1415aed78e63b7105624891c1f
Former-commit-id: 63ed773a74a9b6bb6d21b2f27a8a4af9b1931b4f
Former-commit-id: e21085d28944a04a34584e2b40bd47474f14ef5a
Former-commit-id: 76896a71f79ccc12c7351fcb4d9b709e9b3964a4 [formerly 5907dbdea5d8a95f67e50c3a3041886347f7d45f]
Former-commit-id: 87d7313c664c280f46ba6ff8fd869d61d5bfb7b6