finish removing resources::game_map
The remaining uses are just, checking whether or not the pointer
is null, or resets to its value, or simply storing it in a local
variable.

It might be appropriate to add resources::game_map back as a
function taking no parameters, at least this is flexible
enough to be changed later easily.


Former-commit-id: 73388ffc4b41eb5ce86be35abe77781d20ae9990
Former-commit-id: fb6e8348291db3bd5793c4d25dbc650c8a3f0192
Former-commit-id: 3c75f828d48d262350c0e14b128c9dd8daca161f
Former-commit-id: bd11cf989753de901191ebfbdbc1d9165886232e [formerly a049b76617d4bde8543b7f9dd8b0e781f25c5df6]
Former-commit-id: 78cdf878b368c1d8a71420218712b837194beb90