change saved_game & savefile format
we move code from playcampaign.cpp to saved_game.cpp
we also change the savefile format so that ingame savegames don't
contain [carryover_sides_start] anymore becaus all that was already
merged to both [snapshot] and replay_start.

This commit is likley to break something. If so that will be fixed in
later commits.


Former-commit-id: 869b4c04979382cc9ce0d2ee4329c36a4d8680fb
Former-commit-id: 534910030117a13857d94cd7d70325569a9be4c0
Former-commit-id: ef93d6fd3d00f7311caa5a35bc0ccd6c994e2d1b
Former-commit-id: 9f5f35fec9b5c4e6a2727587a776e46f95486b9d [formerly ec8e1d006dbdacb5b3a933e311e07055f7b1bbe5]
Former-commit-id: a942e893df0dc80a37f5beca3784ebf6b3f75bc9