campaignd: Don't use GUI1 markup in [upload] response
Wesnoth 1.7.x and later use a GUI2 code path to display the [upload]
response message, which means GUI1 markup won't work with it. They do
not enable Pango markup either, and in any case, formatting messages
properly should be the client's responsibility, not the server's.

Perhaps later we should allow including a secondary message in the
response or something like that.


Former-commit-id: b24872c4fffc6c851f9bba9451c8a93dffc12755
Former-commit-id: a809e7b5e9f9c2a0ccea8dcbe068a564768476c5
Former-commit-id: e27251a7f1f987969c76327eef65a4ab68c159f2
Former-commit-id: 7386db9767ad43f83f76ce7ef9244d8ef50ce672 [formerly 873db0e9853b08070f7cdc81d305082c648e7d88]
Former-commit-id: 4eedd02efb4ed96926bf6916db41b4ec6ac2b24a