display takes a pointer to whiteboard manager
It uses a shared_ptr, to prevent crashes during play_controller
initialization. But it only holds a weak_ptr, to preserve
previous behavior of skipping whiteboard ops when the ptr is
NULL.

It requires also for us to construct the whiteboard before
constructing the gui, so that we have a valid pointer to pass.

Most likely all of the managers will have to be moved forward,
in this commit we move forward the whiteboard and pathfinding
module, anticipating future commits.


Former-commit-id: 0584fe46033541e3fd50a969b121bdce4bfbfa2d
Former-commit-id: 782141a70e69d93a416b8258af28acebd1a11601
Former-commit-id: 88a7f1e9b4be9349185dc4882bdf5b8b2e00afab
Former-commit-id: e492affddf7a319c237d0432dc78a724ac759450 [formerly f61045f09a0c8663d1c63dec4e49213d94af9062]
Former-commit-id: 0299ef1ae51ddbb82669a04fb00e0233c66474c2