new_release: Workaround race condition in wesnothd SIGHUP handler
If we try to send a message immediately after signaling SIGHUP to
wesnothd, it won't get caught by wesnothd, probably because the control
socket is being reopened just as the message arrives. Thus, we need a 1
second delay after the signal.

This was fixed in production a while ago but I forgot to commit the
change.


Former-commit-id: e11ac154849d647cc707ac1be38068a9cc0a69bf
Former-commit-id: 779b3e829c38a4aaf204782edf7113af697fe0f0
Former-commit-id: f5a7f73b5d064bfab84ffefaf2f9aef2b90875d9
Former-commit-id: e557348f1b7a1f1fce8c26acf772edb9823086e1 [formerly ce0c4a07b39f61ea7d0a35d26ff9097801b29a37]
Former-commit-id: e8822a279cf8c811d402b7cc28ede6e9750aeb7f