add virtual fcns to display, remove refs to game_display in reports
This commit adds virtual functions covering "observers",
"displayed_unit_hex", and "playing_side", in display:: object.
This allows to avoid dynamic casts in the reports.cpp file, and
to eliminate a header. It avoids the possibility of segfault when
using the editor if the reports context isn't set up correctly.


Former-commit-id: b38f23b8e3d166b707d30ea666c452bab7198bba
Former-commit-id: 7801a36fbc54b3a332dc9dbf4248527e51930c02
Former-commit-id: 63f93f366f61d5e69c3b3e3a320dba2e35b9a251
Former-commit-id: eb42f34df0fd7ccdf1b7e510bf77635c3158769d [formerly 9f10755021cf97c2f7b76c07fa517f140c558931]
Former-commit-id: c43e264313b8f4a87f410f9265a799684f6bf3e8