Suggest checking the data dir path when unable to initialize fonts
Not being able to initialize fonts is usually the direct consequence of
a failure to read data/hardwired/fonts.cfg, possibly because the game
data dir was incorrectly set or detected. Thus, it's the ideal point
during early initialization to provide a hint on stderr for players.

I specifically avoided mentioning anything about "installing" the game
so as to not contribute to the common misconception that the game needs
to be installed to run.


Former-commit-id: cc26c0ee733d971225cc4f7fa92aa3b5962e6302
Former-commit-id: 07652673a8de67c3a19d6a41a8c0ee664a590943
Former-commit-id: cc6f52c21148c312636bdbf338dda3a2de88689f
Former-commit-id: b564354850ffd5013450a52a8fa062b474cb7bec [formerly 7c914dc7f1b38fa0f6cd2ea91480f0cb8ad3be4f]
Former-commit-id: 1af077475dd02f452a4f28aa1d254b6d07bacc79