Suggest checking the data dir path when unable to initialize fonts
Not being able to initialize fonts is usually the direct consequence of
a failure to read data/hardwired/fonts.cfg, possibly because the game
data dir was incorrectly set or detected. Thus, it's the ideal point
during early initialization to provide a hint on stderr for players.

I specifically avoided mentioning anything about "installing" the game
so as to not contribute to the common misconception that the game needs
to be installed to run.


Former-commit-id: 25335afbf4fb8bf4efb4b64faf065d32b0378d2a
Former-commit-id: 85b3f04ec6d08d2fdb9007fd23a00cc2cb471fbe
Former-commit-id: 8a201c2babff4d513beec647cbe8e32a6e640225
Former-commit-id: bc95cc158579d4a55017e048f0248836aa087d21 [formerly da2e2266f83e14065929de4e203ef2ec391c3268]
Former-commit-id: 784cf1af8829ed1250b289e3908a3d794c4394af