optimize parser
According to valgrind, the tokenizer::next_token function is called
several million times during a typical unit test. This commit
changes the parser to allocate the tokenizer as a member variable
of the class (on the stack) rather than on the heap using new,
thus it eliminates millions of calls to new.


Former-commit-id: 5758e849ae2a4a12782f8a1ef2c5ee009c862b07
Former-commit-id: b22206360a482f136b47494f59df30b41c9aae52
Former-commit-id: 96feaff496245fd0b956d7fb8346a38adffdf76d
Former-commit-id: 8b757b03cf28eec1f719e5aed136130822e8714c [formerly 87ef52f8bb05c617174f70533378ac4e443490eb]
Former-commit-id: 8b3e9018cb3971d863155b70941fe55436af0af9