Call set_button_state() from playsingle_controller::init_gui()
Fixes the minimap buttons sans the minimap unit/village drawing pair
being in an indeterminate overlay-less (blank) state for the duration of
the initial 'start' WML event in scenarios, unless they are interacted
with first.

set_button_state() has to be called after play_controller::init_gui()
finishes because that method calls game_display::begin() game first,
which in turn instantiates the theme UI GUI1 buttons. Notice that
set_button_state() is also called when the End Turn button is enabled
(e.g. after 'start' is finished).

The reason I'm not calling this from play_controller::init_gui()
directly is that I am concerned about possible interactions with the
replay_controller, which currently seems to do even sloppier work of the
GUI before the end of the 'start' event. I might move this call there if
that issue is sorted out, since it'll be needed for the
replay_controller anyway.


Former-commit-id: 75ae7f420bb918618f121e9af1c448df2624ffcd
Former-commit-id: f7b893206334c283d8c8a1540c662a72f3cb7a81
Former-commit-id: 0b2e0c190007c5cca3475c2e625a83dbce58149d
Former-commit-id: ff5e18645230f0ba1b67a0286db8be9a8f81b949 [formerly 4d8ec0967aff5a7595554969a2dbf8ba1335a487]
Former-commit-id: a0b97154b9216b4a44a00372a911174c7323ea0c