Call set_button_state() from playsingle_controller::init_gui()
Fixes the minimap buttons sans the minimap unit/village drawing pair
being in an indeterminate overlay-less (blank) state for the duration of
the initial 'start' WML event in scenarios, unless they are interacted
with first.

set_button_state() has to be called after play_controller::init_gui()
finishes because that method calls game_display::begin() game first,
which in turn instantiates the theme UI GUI1 buttons. Notice that
set_button_state() is also called when the End Turn button is enabled
(e.g. after 'start' is finished).

The reason I'm not calling this from play_controller::init_gui()
directly is that I am concerned about possible interactions with the
replay_controller, which currently seems to do even sloppier work of the
GUI before the end of the 'start' event. I might move this call there if
that issue is sorted out, since it'll be needed for the
replay_controller anyway.


Former-commit-id: 50fd403b074cfc2e5212939bfba847d44801dcd1
Former-commit-id: eef43f20c080bc1fcb8dc4c81105600c6e9e99b3
Former-commit-id: 6998abfc75d4d9f46288463be300058cdfa8cf0d
Former-commit-id: 5eff8bcf4c395aa1c05f06d1d0548a81f5c1e1fd [formerly b68d6dd628460668d9a2a842930e865645ec3341]
Former-commit-id: a0ca02eac56101248b38cda2066bde971fbc6e50