campaignd: Replace vector<pair<k,v>> with map<k,v> for the handlers table
std::vector<std::pair<k,v>> has worse look-up performance than
std::map<k,v> in general, not that the difference is too noticeable for
such a small dataset like campaignd's handlers table, that is built in
such a way its items are arranged in descending order of usage
frequency.

It still feels like this is the best thing to do to keep future
maintainers sane.


Former-commit-id: 2dd38be94b9964a2a9011767927dfd28c392ede6
Former-commit-id: 1e5cf41521a41f391a79bd20959fd4cbf674da03
Former-commit-id: 655cf36e54dc67f5720935ddadba8673c65c119a
Former-commit-id: 7aaf52412f5567cc3e88040f94dd2231a60622f6 [formerly c8e5c545646555622656ed9a8f465d33db4d6518]
Former-commit-id: 4df8733a3d936880912221500c4add6492e5ab42