fix random map generation
this was previously broken because we passed a temporary pointer to
generate_map. While fixing this i moved it to saved_game.cpp

i also added a call to expand_random_scenario in mp_game_utils so that
in a mp campaign players advance to the next scenario which is randomly
generated. they have all the same scenario.


Former-commit-id: 246d57203bd49fcb2e8b55ec5b03b2b55f9be42b
Former-commit-id: 716bd6d109787e960d6e83e57f09457c50cf9cbe
Former-commit-id: 2a644e834523e3dff8c78eb0d3a27dd087b76be0
Former-commit-id: 7ce7815bdb5e710ca12cf2ccee5ff951e4e192e0 [formerly ec3b64ef533404d87bcc7997210f709dbd7c8d34]
Former-commit-id: 5b5177ed42c22217fb8d683bd91f7333b4bc23cb