fix get_user_choice for other sides
as the comment in replay.cpp says we want "send data to others" so
calling "pull_remote_user_input" is not really correct. Since
plamp_controller::handle_generaic_command("ai_user_input")
also invokes send_data we just don't call it with "sync_network" in this
case. This solution is still not optmial but it has the smallest chance
to break things since epecialy because this should go to 1.12 too.

Conflicts:
	src/ai/manager.cpp

Former-commit-id: deeb78647bd1e21a383026e1d5bc4fcc942ed719
Former-commit-id: c361bde3f7a95abd4a87e6bdf3b866260261c827
Former-commit-id: c705ee470e9fa1fd2a39ec00a8ed2684fca64d96
Former-commit-id: d2c0002f9ba5a06e6f90b63e9794b7f52bfafc5a [formerly 76a249ca2c6a3c0c4d52cf628e4f5c7719065964]
Former-commit-id: 9e29cd0cb39a8fb52a3ff0f9cdee492bff190e8b