add fake_unit_manager, split off from game_display object
This commit adds two classes -- fake_unit, and fake_unit_manager,
both split off from code contained in game_display. The display
object is reconfigured to hold a pointer to the manager and
display the units it contains. The rest of the code is configured
to add fake units to the manager, not the display.

This improves encapsulation and helps to reduce the game_display
class.


Former-commit-id: 0bc8ac73b8d99401c3c204b3725de7b4fa9d1ce6
Former-commit-id: 4fb25c14c82c40c8f470754d76c0a7b9849df855
Former-commit-id: cccaac12b7ad93397247bd79c7550fb5fbbefa6b
Former-commit-id: 57e20cd97635bc043c657c8f6fe95bb5a0dec376 [formerly 5f3b22324eea6581daa42309a987edd1030ad4d0]
Former-commit-id: 1fcd81e00bca71a8293f826e717b907e0affc273