use UnitPtr (intrusive_ptr) in unit_animations
In this commit the unit * in unit_animation is changed to a smart
pointer. This is to prevent the unit_map or some other object
from destroying the unit and creating a dangling pointer
(reported to cause segfaults).

In this commit, we also change the fake_unit class to make sure to
increment the refcounter when the fake_unit is constructed. The
reason is that the fake_unit holder is supposed to own that unit
and let it expire when it goes out of scope. No smart pointer is
supposed to delete it.

This mechanism should possibly be revisited, for example the
fake units could also be held in a reference counted pointer.
With the current system it is still possible that the fake_unit
will go out of scope and be destoryed while the animation is still
alive, this change only prevents smart pointers from destroying it
before it goes out of scope.

I also comment out the reentry_preventer class uses in
unit_animation.cpp, to prevent assertions from happening when I
test for example with the "double debug kill" version of the bug.
The reentry_preventer seems to be unnecessary now.

This addresses bug #18921 and maybe others.


Former-commit-id: 3e2c0724a44fa5db3d796017aa2e13a58ac45cb8
Former-commit-id: eeeba275f5347b36a8af8b0d6e0746b424f5f52f
Former-commit-id: 52dfec8899ad7c503b619d6852755951449bc2be
Former-commit-id: d6052fb3adb388f4da5e65764fb39ce6e9b830e2 [formerly 9de67a4bb5bbef32099f99b6866cf2f6fe6d1731]
Former-commit-id: 01f2fa1cf6949565070eab3dac76b5303f68d986