add recall list manager
This commit adds a dedicated recall list manager class.

The purpose of this is to

- Simplify the code that interacts with the recall list. Prior to
the commit most such code was based on iteration with explicit
iterators, and called global helper functions implemented in
unit.cpp to wrap the code that finds a unit in a vector. It turns
out that interacting with the recall list was the *only* use of
that code, so we make it a member function of the recall list
manager and take it out of unit.cpp.

Most of the code that touches the recall list was previously
7 or 8 lines with a for loop, now it tends to be 1 or 2 lines,
although further refactor may be possible.

- Improve encapsulation. This makes it possible to track how
other classes are interacting with the recall list, and may
make it easier to debug recall list problems by adding debugging
output to the class.


Former-commit-id: 947623fc9fb9a3fd0efd7c5a4a7b63f497c4af7e
Former-commit-id: 16023f8b2aa9f901f25c6230130a78b3b1ad2689
Former-commit-id: 0d5debe229060a8413b93d909b432a41d209f02f
Former-commit-id: e8b2dcceca1d86418e904e51c48e1ff3a675894c [formerly 6df1437f21784c6e3356a0e3cb3c15e7dd88f52b]
Former-commit-id: f70ce717a7295948bb1a391ed45b25db3823e4db