set travis back to gcc 4.6 version
The gcc-4.8 patch requires travis to download and install gcc-4.8
from an ubuntu ppa. This takes a few mintues and the gcc build is
already fairly close to the timelimit, so this change can push
some builds over the edge. Since it doesn't seem to be strictly
necessary right now, I comment out the code for v 4.8 install.


Former-commit-id: 5d688ed20b845601a6000bd7d908d707ae167834
Former-commit-id: d37628b7becbe2cc1f5bbde3b334b322e8fa04e6
Former-commit-id: 2abbd6bb6114b94d72d0d37f6a2e277ed0cd22e7
Former-commit-id: c27962d90645860d8110456fad100206a1f454dd [formerly 8dd25d9f126ba1104f44ad2f252a5c15918cdd27]
Former-commit-id: 9b67d229a90b395f116deb06a015bc633b62386e