disable all tests on gcc builds, to prevent timeouts
It seems that we cannot avoid the segfault without adding gcc-4.8
download and install to the test sequence, but this seems to push
us over the time limit consistently. In this commit, go back to
using gcc-4.8 but disable everything but the wesnoth target itself
for the gcc build to save time.


Former-commit-id: 67061c7b017d082e22299b1e5da409b20d620604
Former-commit-id: 8fcd0089721868f6603714350d4c75c4d98b7b20
Former-commit-id: bd171c7faeba383627d1a89cd12508e70d879089
Former-commit-id: b7771df9957f7f3fa0d9e378176230d03c10115f [formerly f130034c5adeb4d44f4b5e5f5da033d48f8769f0]
Former-commit-id: 3e5b15cc2f8d5752249ee3162708684bcb28b5a2