fixup 6a839a6714cadfaa05101c9e6dae62be02ded118 [formerly 4ac8015fcaa1c0cd73a5463d051a988aa1ad34b2] [formerly 10dd54d4103704e687dc88d44101be2dac380e52 [formerly 1d7c9bb010b80aabe76176b3e1ea2a3c21b41d86]] [formerly 7447f0b32cf4c8433777fdd31e219654eb04ba4d] [formerly 0ef5af302094717bb17720340fdfc12fe8bafe95] [formerly acbaacf83affe5cfeffba2706105298cf8bd8630]
In this commit, the order in which fake units vs real units was
changed without realizing the significance. Comments near to the
fake_units code makes it clear that the fake units are intended
to be drawn after the real units, so that e.g. when the whiteboard
is used, the fake units draw over the real ones.


Former-commit-id: be4af8d420c0d3d90e6881c3fa2f6ecec13313c6
Former-commit-id: 83ee163a14324dad8212af5f798df937b1d5fc02
Former-commit-id: 16605147bc54687b1c0898a12d2a106718f0cd5b
Former-commit-id: 04623de389eb733c1c1c5e040eedbcd5eea6c3ad [formerly 0d0c3690fcec1435acc789351f5902b6a62f4ca0]
Former-commit-id: e7c367c8f3068fc8da16bde98612351398773d9e