change fake_unit class to fake_unit_ptr
This is consistent with the introduction of the UnitPtr class.
fake_units really aren't different from units, the only difference
is their life time / allocation and ownership. Since we are trying
to use reference counting for all units (to make them safe to use
with animations), the fake units need to be managed by a reference
counted pointer also. This is the easiest way to achieve that.

I also remove some odd code the [move_units_fake] handler --
there was explicit code to remove the fake units from the fake
unit manager, but this is redundant as it is the responsibility
of the destructor.

Code Blocks and VC project files are updated, but
Code::Blocks Scons and Xcode are not.

Former-commit-id: dbafe17e07ac138a88da300c981f7384b1ef94ff
Former-commit-id: c5c2d1ca8ff25700f488efb4e97f5bb4930b5b6f
Former-commit-id: b4d06cbe6e2774499a9e4af9c5a02a6b2c9e77c5
Former-commit-id: 06150e7c1a4733a21c63a874cd279b0f6fbc3ced [formerly 95375229f00cd07970caa4bd4306ae224a7c0935]
Former-commit-id: a3ae7ac1fc37da457be1d4e97ae7a8acfb29bb8b