Apply local ToD lighting to hex overlays when applicable (bug #22215)
This replaces map-wide ToD lighting with local ToD lighting for overlays
associated with locations that have custom local ToD lighting. The
method employed for this is more or less the same already used by
display::get_terrain_images() bar the complications associated to
dealing with terrain transitions, since overlays don't have those.

Note that for this to have an effect, local ToD lighting must be enabled
in Preferences -> Advanced. This is the default since Wesnoth 1.10.x.


Former-commit-id: 6c5be7032e739010fcc3300af0a19b91a5349086
Former-commit-id: c145ebeaaa924580c11d350433656688212df0b3
Former-commit-id: 383da486abb3332920afb38f628524b4cfc09f04
Former-commit-id: ad7259593112376c06f361c6cf6f94763cc54637 [formerly bfe58294b220357765731ea40f3ef99036d08fc3]
Former-commit-id: 4bbef355b59b99c62c5b071f5307cc8c357bc418