Apply local ToD lighting to hex overlays when applicable (bug #22215)
This replaces map-wide ToD lighting with local ToD lighting for overlays
associated with locations that have custom local ToD lighting. The
method employed for this is more or less the same already used by
display::get_terrain_images() bar the complications associated to
dealing with terrain transitions, since overlays don't have those.

Note that for this to have an effect, local ToD lighting must be enabled
in Preferences -> Advanced. This is the default since Wesnoth 1.10.x.

This patch is more noisy than I expected, so it might require some
additional testing before landing in 1.12 to ensure it does not
introduce any regressions.


Former-commit-id: 7f06236be28fa8cd9d73ec0f668217a48771a606
Former-commit-id: de0e66267f870384956d7746edb762f67f6b67f5
Former-commit-id: 5cadfb0aa5650c20bc63cb93d44c3ed29376ee4d
Former-commit-id: 33dd220fd8e328d88ab71ad6a9ad5d5da950cafe [formerly f1a60e5ab76af22c8c3703f671eb6fb68c7ddf9f]
Former-commit-id: e896d2c78a80ffc1c4dabe558787a28ec1b3e028