remove mp_game_settings::share_view/maps
This is a leftover from 1.0 where this was adjustable in a dialog. This
is not the case anymore, the current behavior results in
[side]share_view= always being true in mp which just means we have one
less option for mp scenario designers.
Even worse, the previous behavior results in share_view always being
false for scenarios in mp campaign that are not the first scenario,
regardless of the settings in the scenario.


Former-commit-id: ee277e60d2b3340f93b7a9fe67a8ee35b4fd1739
Former-commit-id: 360845b32d8a4ab1f081fd55d032d62d3942895e
Former-commit-id: 14352c2a052cf503bbe20e45810f6c82a7a3e9f5
Former-commit-id: 1a3c4de825dc11a2023d5712839db6bf20f00ab5 [formerly 63e8159f6c01f83188e0446fb4ac7dc4578e6a70]
Former-commit-id: 1c5ace7ff34cace8966234cf5d94edc1951e1383