campaignd: Use find() to locate a request handler
Makes it so for each request received we use
request_handlers_table::find() instead of a plain linear search to find
the handler, using the tag name of the first WML child node we can find,
instead of trying to locate one with a known handler id.

In theory this allows taking advantage of the request_handlers_table
type's (currently a std::map) own lookup mechanism.


Former-commit-id: 563f9aa3580f2d5ae358acf893a34da07c08bd70
Former-commit-id: 220a659c2f389f40559fec4034d62d136e9c8bfd
Former-commit-id: d6733eecfad3abaf254d1048e6bd5d0a989dd72b
Former-commit-id: 1b5d0ea30896096d1c52adcc2d14ce6810bef7b7 [formerly 12ff7111de2b0ff4c07768c0ee32c37a86e5ce8c]
Former-commit-id: 96a6ac91139a9cd9803e9df4a17ce247c723c7a2