fixup 8e6d4423dab136cf5009c54d30f9aefba0a81c08 [formerly a3ae7ac1fc37da457be1d4e97ae7a8acfb29bb8b] [formerly 06150e7c1a4733a21c63a874cd279b0f6fbc3ced [formerly 95375229f00cd07970caa4bd4306ae224a7c0935]] [formerly b4d06cbe6e2774499a9e4af9c5a02a6b2c9e77c5] [formerly c5c2d1ca8ff25700f488efb4e97f5bb4930b5b6f] [formerly dbafe17e07ac138a88da300c981f7384b1ef94ff]
There was a small mistake here: The unit constructor
which takes a unit type first and 3 other parameters,
has the 4th one being an optional gender, and the
third a bool. fake_unit had a constructor mapping to
this, which sets the third to false. I over looked
this when factoring out fake_unit, but it only
generates a warning to cast enum to bool on MSVC
compilers.


Former-commit-id: 0dadf7171104d4a7421e8004f870ade846ce7f6b
Former-commit-id: c25dc20d45d37c7524402b0a0233a2ee125e3c00
Former-commit-id: 7422764f1be6197cf8dae45db519532ea6a32c57
Former-commit-id: 59182fe6cdc9f7ac9fa6de0bea50f3358ca7ce5e [formerly 9193063c459ad05f49394c7c3a813b70651d11bc]
Former-commit-id: f4d10ef96de6e2110e36c06c73ddfe01e6afc368