Fix the linking error found in the last commit.
The error was caused by forgetting to add the new file to the
CMakeList.txt file.  It was then found that the new file did not
compile because of non matching variable names, those were changed
so they match.  Lastly a comment asking for reveiw/help finding the
issue was removed from unit_types.cpp as it is no longer an issue.


Former-commit-id: f41ac3dc02970437d15f46a3d1bec31e08a27f2e
Former-commit-id: 2647431a796326e58080f41e776a1494ed71ee70
Former-commit-id: c81f56923af8b28f852696a22058a5a44ebab2a6
Former-commit-id: ba16456f2b89bec9fc2ebe7192841d00c9512b69 [formerly 482aa0ff8ece2a93dbe52e04d87e2cc0c20e3582]
Former-commit-id: b0c0fcf2444fa77da5b9906f574ff3192376ee5e