add display chat manager, notifications support outside game_display
The functionality of tracking observers and displaying chat messages
is moved to a manager class, which the gui owns.

The functionality of displaying notifications is similarly moved out
of the game_display and to a private namespace. (Static singleton
pattern seems okay here since there really won't need to be more
than one of these for a single application, it seems.)


Former-commit-id: e228b42e924746bd6f891f7e69d64e9af8c46f82
Former-commit-id: 9c76344bab4549e7b7474bd707b60471d184c152
Former-commit-id: ce55d826ad5c11d430a557e2c63d6c663d878fb0
Former-commit-id: 3c1d50e8bf4e25678303c763ebd52059a6ea3e42 [formerly e458764c20e7b29e710e86e6ea328bb2b875fa5d]
Former-commit-id: eaf7516e73c98a604ef2175ab21fcefa27b2a461