add wmi pager class, to allow unlimited wml menu items
The pager is held by the play_controller and wraps around the
wmi_container, if there are too many items it adds "More Items"
or "Earlier Items" entries. These entries have special reserved
ids, if they are selected the pager traps them.

Future work might be to allow the page_size_ to be selected in
the preferences, that is currently left as a variable to make
it easy to support this.


Former-commit-id: cd7e316bd6fa7cc883bc8d8cbea0a0e91ae03755
Former-commit-id: f55f76da3aa63da4b17a51cd0a4a67577a142802
Former-commit-id: 125a9ee4e00852de0a5dd766f30d7cfc769f4000
Former-commit-id: 88fad341a153880e3a1c71d8fce4c5eda18cea84 [formerly 8b2dca205ac6fe45ab1d2c541e2d4b37828f2f19]
Former-commit-id: 48f7a77ee03437f85aedb418d279eed8347a2574