Demote "skipping textdomain" error to warning level
The "no location found for '<path>', skipping textdomain" error message
is usually inoffensive and merely means that there is a textdomain
declaration pointing to a directory that doesn't exist, so it shouldn't
really be an error. Since we don't have anything midway between warning
and info messages, this should remain in place as a warning for WML
coders typoing textdomain declarations or forgetting to include their
add-on's translations with it.


Former-commit-id: 0d80d3c7c50ccaa57ec7de868d38fe0a785a149c
Former-commit-id: e7ef86d9f325705b9334ba281b21a321d5a8a1b8
Former-commit-id: 71f23ffb1920edff7a38bb123cce0296e8026f95
Former-commit-id: 61e920900b1e63ff213b650ce8ff120f69f80a72 [formerly 5c4d8fda40157f518fbc2207d51710ee6140ef6a]
Former-commit-id: 4f3e2715d9c4cb5a7a8657b770662377e55d1e1f