Demote "skipping textdomain" error to warning level
The "no location found for '<path>', skipping textdomain" error message
is usually inoffensive and merely means that there is a textdomain
declaration pointing to a directory that doesn't exist, so it shouldn't
really be an error. Since we don't have anything midway between warning
and info messages, this should remain in place as a warning for WML
coders typoing textdomain declarations or forgetting to include their
add-on's translations with it.


Former-commit-id: 791d3212cdfe707b6ceec94c46ee3b953903ae0a
Former-commit-id: 3ff27a0ecaed6b67e62221187422535ac009a370
Former-commit-id: 9319f093e28c187e22822a8d3a68c12f2180ce94
Former-commit-id: 07a26731e17238d8ed4baaf5e559a473ea283058 [formerly 7c6ee0798019c4b49f1895d7a15def80a9b0c6d2]
Former-commit-id: aba57741febdf9e05f8726ebf449910c7aee343a