campaignd: Log messages to a new log domain of our own
Until now, campaignd logged its diagnostics to the network log domain
with the show_names parameter (of logger::operator()()) set to false so
as to avoid showing the log level or domain names in stderr. There are
two issues with this approach:

 * In order to tell the difference between informational and error
   messages, the messages themselves need to include a severity label in
   some form, which defeats half the point of using our standard logging
   facilities (though not the other half, which is having timestamps).
 * In the event that messages from other domains appear in stderr (which
   can only happen if their severity is 'warning' or more at this time,
   since campaignd doesn't change the default severity configuration),
   the log could become a jumbled mess.

Note that wesnothd (which produces a larger volume of diagnostics all
the time) does already use a log domain of its own with no overidden
show_names option.

Giving campaignd its own log domain should allow us to add more optional
diagnostics later if needed, and right now it allows us to do away with
the unwieldy severity labels in the log messages themselves in favor of
choosing the right logger in the invocation -- although this last point
is beyond the scope of this introductory commit.

This commit keeps our previous logging macro, LOG_CS, just replacing the
log domain and severity (info instead of err) and dropping the
overridden show_names value. Later I'll take care of choosing more
relevant severities for different messages. For now, we do like wesnothd
and select the info severity as the default during initialization so the
messages can still display.


Former-commit-id: 02e1d416cd9b8b10405191eed84bbdd2eeb5060f
Former-commit-id: 543f90246fead48b13ce7c6741ae31463fc32841
Former-commit-id: 2fcc20ddcd0b453fc3a37d60311bb63d973ecc10
Former-commit-id: 62fb04b001431fe04aa9d23d0b0cf4f4ad9afd64 [formerly 205eb12feac9f1478543496ffd0c4a3efd2a2b3a]
Former-commit-id: 2feb49759334e8e6bbe7fc351d4e231190d99a68