fix a bug which could cause the terrain_builder to crash
The terrain builder logic assumes that the min_types list is
empty whenever the min_constraint pointer is null, and based on
this, dereferences the pointer. However, in subsequent passes of
the outer loop the pointer is initialized to null while the list
is not cleared. According to coverity this can actually cause
a null pointer dereference in explicit cases.

We fix it by explicitly re-initializing the list on each pass.


Former-commit-id: 761de74c7a4e72b9f53d1dcb507e0965f5c769e4
Former-commit-id: b9066384d44b5decf4390dd468437cbdf4c7d974
Former-commit-id: b207a0207c809bd18bf22076209a029d4cab5193
Former-commit-id: bdb7952c6e03cd61b44dfb168e8c2ee0f72f49d8 [formerly a8eba023f94da2cfc6473bab7a7eaaf45bc6237a]
Former-commit-id: 5be2d72eeb9b6c2c00d9ca4bf1fab1e629be9f9f