fix a bug which could cause the terrain_builder to crash
The terrain builder logic assumes that the min_types list is
empty whenever the min_constraint pointer is null, and based on
this, dereferences the pointer. However, in subsequent passes of
the outer loop the pointer is initialized to null while the list
is not cleared. According to coverity this can actually cause
a null pointer dereference in explicit cases.

We fix it by explicitly re-initializing the list on each pass.


Former-commit-id: ec96a6c6bb3a5709efc4124fe303a01a048003ac
Former-commit-id: 007be2c677686e836275d18473caf4bc8931ae08
Former-commit-id: a2423a6fc29d6476ccdc7f0739a433d60c05a6f6
Former-commit-id: bd01cd8897ea72994c66795375d77226896ead7d [formerly 0b7c2eae1b28a4b8f246ba577872feeb04122799]
Former-commit-id: d71276f473e17aed906539a2a52e0df87470e32a