fixup inefficient invisibility calculation for units
Units have a function with signature
bool unit::invisible(const map_location & loc, bool see_all)

which determines whether the unit, if it is invisible, would be
revealed if it were moved to map_location & loc.

The implementation of this function was needlessly inefficient, it
requires us to loop through the entire unit_map (all of the units)
to determine if any of them could ambush the unit.

However, in the course of this check, it filters out any units
whose location is not adjacent to the target location.

This function is called constantly whenever the display is updated
or the minimap is drawn, or when a unit is moved.

The improved implementation does not loop over all units, instead
it gets the list of (6) map locations which could be adjacent and
have a unit, and then queries for a unit at each of those
locations.


Former-commit-id: 1f2e2f1a18b204563ee0c73f5b21c1ded4fe2706
Former-commit-id: 6bdb3195156d49482f77c70420f8c12019d52a3a
Former-commit-id: c8e0614b3d5ca769efc37152213be61cf69fe0ff
Former-commit-id: 3d87863e5dac734c237e2f721ee518302abf7ad1 [formerly 3c34a6cfae9816bd73ce3837e06be91de560ed66]
Former-commit-id: 37d29be4c687610ecd8c9919fc6543ee0f221d6f