fixup inefficient invisibility calculation for units
Units have a function with signature
bool unit::invisible(const map_location & loc, bool see_all)

which determines whether the unit, if it is invisible, would be
revealed if it were moved to map_location & loc.

The implementation of this function was needlessly inefficient, it
requires us to loop through the entire unit_map (all of the units)
to determine if any of them could ambush the unit.

However, in the course of this check, it filters out any units
whose location is not adjacent to the target location.

This function is called constantly whenever the display is updated
or the minimap is drawn, or when a unit is moved.

The improved implementation does not loop over all units, instead
it gets the list of (6) map locations which could be adjacent and
have a unit, and then queries for a unit at each of those
locations.


Former-commit-id: 792d94d714d8f2e586e72ce438607b653e8e3785
Former-commit-id: dfb76bba6ce383ac83679289b58d482a1aa0c38f
Former-commit-id: df17afd6f847471a9c00a65722f15b038183bf55
Former-commit-id: 4ac368cf9a49cd46ffe7c19accd5d3cd9efac590 [formerly db596262fe0c7df8f179964d1d3df8c2e10a5f17]
Former-commit-id: 1c16acb43d1ff04511d97de5e5b815799f899052