make the "alternate" clang build use -O0 optimizations
The alternate clang build currently uses strict compilation = false,
this was requested so that we get as many error messages as possible.

We would also like to have an -O0 build run against the unit tests,
as sometimes -O0 will segfault while -O2 won't, etc. So in this
commit the alternate build uses that option also. I also rename
the environment variables, so that the build matrix is based on
casing out whether "alternate_configuration" is true (not strict_
compilation) as previously, and the alternate_configuration flag
leads to options being exported in the before_install phase.


Former-commit-id: df38a4f45541f64d6b8372b11636e3aba4aafbec
Former-commit-id: a405c9a3b886a59f2d5ed87a4a0f04a7d7beeb91
Former-commit-id: 3cbe3b12422b21e6b78f42ec98efff2cdc1d157f
Former-commit-id: 29a337dd59e941c241fdf1038673646c5868019f [formerly db1885ee3bbd827b382c8aaa3fbae39b81b0cace]
Former-commit-id: 3c1a91ca034957c0f10b48d912a32e30a991d01f