add lazy parsing of comma separated lists to unit_filter
The old code seemed to desperately want to avoid calling split
on large lists of strings repeatedly, however it did this again
and again for every unit. The new implementation remembers the old
splits for the next time the filter is applied, but makes sure
only to split them if the old code would have split them.


Former-commit-id: 8c06bcb93e4da2dc9cea0fe79fff65fe97688946
Former-commit-id: aa724d9fb0051af48a10f06d1faa512cbe9aa792
Former-commit-id: 1317342a75ffb4b3ed2c036b4a03a79754463422
Former-commit-id: dcf9bda214ee18a2997a85fae9650920964fd9f3 [formerly 39f75c552d6462fc45f06d86efedf58fa995cefe]
Former-commit-id: 3d8d603566cb1f0d9de63c4fafda467a75406d93