addon/mg: Include Update All on all views and enable it when applicable
This makes it so the Update All button appears on all views, disabled by
default unless there is at least one ADDON_INSTALLED_UPGRADABLE add-on.

The rationale for this is that with the Update Add-ons shortcut in the
Connect dialog gone, the possibility of upgrading installed add-ons is
not as visible as it was in versions 1.5.3 through 1.10.x. Making the
button visible at all times should hint at the existence of this
functionality.


Former-commit-id: 20a784f227c3a2aa34b6b7cbca3f45c5ead9f895
Former-commit-id: e8f8e6642381ff303c76165e146c9114fa3eeebb
Former-commit-id: 8b8b39fce5f5420daadb225750fd2753ec893869
Former-commit-id: b5ef32a12b6051592225716a110b9d00c77e1969 [formerly e755e614c92d1c4e51444b957cef6c6b03c05bf2]
Former-commit-id: fd51e8356dd976a5941b138e443533e022188acf