addon/mg: Include Update All on all views and enable it when applicable
This makes it so the Update All button appears on all views, disabled by
default unless there is at least one ADDON_INSTALLED_UPGRADABLE add-on.

The rationale for this is that with the Update Add-ons shortcut in the
Connect dialog gone, the possibility of upgrading installed add-ons is
not as visible as it was in versions 1.5.3 through 1.10.x. Making the
button visible at all times should hint at the existence of this
functionality.


Former-commit-id: c8949ecadd06f2d3f32af27e9191aed8b72544a7
Former-commit-id: fc8ed0fcc01eb88af4cd1261caabace4d5a04a98
Former-commit-id: a24796bd120dcc17efdc8e4d7912c8212bc4d9e9
Former-commit-id: cba1349ec94a18078b2f5889d6483e26641a06cd [formerly a868b5b30355d7cbe5b294ef0b2fd9b70ef0ac6f]
Former-commit-id: 9cdb51ad5d3a95ec19973ebaae6653390a55a5f9