addon/mg: Choose "update" vs. "install" terms when using Update All too
Previously the "update" terminology was only used when the Upgradable
view was the active view. Now that the Update All action may be
available on other views, we need to take care to use the correct terms
for it too.

This commit incurs in no changes to existing translatable strings.


Former-commit-id: 1c6df73bb4b0eb9798c3a1a1666dc9ae1aa6f0c5
Former-commit-id: aeac561cbaa495020b9871d302142bce0e2072e9
Former-commit-id: 1a9886241ab22426ec8af59a5c2fc70ab9c96c80
Former-commit-id: 34ea3420641c9197754d5403325b5fedda647818 [formerly b18f339a1702e350019ab82be84ec6a5d6033d73]
Former-commit-id: 1db78fde18ea901e42d3da0a002c1a4889ae5720