addon/mg: Choose "update" vs. "install" terms when using Update All too
Previously the "update" terminology was only used when the Upgradable
view was the active view. Now that the Update All action may be
available on other views, we need to take care to use the correct terms
for it too.

This commit incurs in no changes to existing translatable strings.


Former-commit-id: d5c21b184bd3395b2d0c0232a767802cc56bda51
Former-commit-id: b147d70650e3bd6f90c249e9c8a49ac7aa7ccac3
Former-commit-id: 02bb9974a81b7fc02072b6689f63c1bf753a27da
Former-commit-id: 2821050f4629b146e8d7331e3c2086f5848c12aa [formerly bd9976191d41c56ba56e2b7f8c51be9cb6323898]
Former-commit-id: 1bd948a0058801a9757cebf958ac4c6001dcddc6