bugfix for #22285
I haven't decided what the ideal solution is for this yet, might
want to rearrange the play_controller init sequence. But for now,
if the game display is null and we need to make an animation, we
use the game state as a filter context rather than the display,
which for all intents and purposes in current code with be the
same. This looks to fix the bug.


Former-commit-id: f2a15f39999b33ab38e2ecfb3d96bced5057184c
Former-commit-id: f88c514226260accccd6aecbbb9364a400f24c89
Former-commit-id: 76d05812ac3da9ae301c4b84f03144bf3d527dc0
Former-commit-id: b83753b5df294ab14fa6165af906ad8839e89113 [formerly 13d362e7da85cd7981592e1fd971a1c05f08dbaf]
Former-commit-id: fa6885953f2cd714230c565e00dc9e5c92a15d10