add [has_ally],[has_enemy] tags for standard side filter
Unlike [allied_with] and [enemy_of], these will match if only if
there is *any* ally or enemy matching the filter. The other versions
of the tag match only if all sides matching the filter are
allied / opposed.

This commit also adds a unit test that checks that it works as
expected.


Former-commit-id: 30aeee7a387256761177ee5012fc89cb8a8271d0
Former-commit-id: 3e62954336a8c5fb0d7237923212e67390b140e4
Former-commit-id: 47a53e80507836d626f8b491361ffc4298a5d0ee
Former-commit-id: 2385934afca1b7e09a443cbb743b1eae5bab8635 [formerly 2a4675f98ea5296dbd42fc735bef05c48ae89b3b]
Former-commit-id: c17dcd1eb591070d4e95818778fbc0bfcfee2250