Simplify --resolution error handling
Reuse the boost::program_options::error class (which we already catch in
main()) instead of throwing a std::string, and simplify the throwing
logic accordingly.

There is no need to provide incredibly verbose error messages since the
user is expected to be able to read --help or the man page to get here
in the first place.


Former-commit-id: 9842acde5bbf32deaa0dfd492ce207b74e82c5b2
Former-commit-id: 97fdd3cf94ef14fb28d5d84919a5eaa776691e34
Former-commit-id: 12b8a622abc31f9161668378fc2a44329c0d8c00
Former-commit-id: 8fe01c5222c45aa266d5b3ab03998cf5152ab8ef [formerly 4eb645fa2c0e318cbdd43da1f52b1efc0c789d34]
Former-commit-id: 1428783a68d9f527c5e66af9a8db59df6f2a4cab