Simplify --resolution error handling
Reuse the boost::program_options::error class (which we already catch in
main()) instead of throwing a std::string, and simplify the throwing
logic accordingly.

There is no need to provide incredibly verbose error messages since the
user is expected to be able to read --help or the man page to get here
in the first place.


Former-commit-id: 13148c99b197c72dbcf33247ccbcffe9e3d742d9
Former-commit-id: 2f141ae49e76293c3cb6e8aa7287f4bf4ad95c07
Former-commit-id: fd5368f52cefe611c4d74a8541c0739e3b6d1d92
Former-commit-id: 335ab1e6387bdf5f546fbf1916f062ae5a147917 [formerly 9c17104006313072fd0fc5bd5cf153988e16d386]
Former-commit-id: ee831c599509f5c98ef4010eb1dc10324f98efa4