Simplify tuple command line arguments error handling
Reuse the boost::program_options::error class (which we already catch in
main()) instead of throwing a std::string, and simplify the throwing
logic accordingly.

There is no need to provide incredibly verbose error messages since the
user is expected to be able to read --help or the man page to get here
in the first place.


Former-commit-id: 6fb3daf0b68a20861f4f000149051b7448cdf3b5
Former-commit-id: ccdf1d2c94d892739e6ccb461a6a8cde5bb7e601
Former-commit-id: 3ab09bb74c9a3e9337c2c7a40475ed52061cb745
Former-commit-id: 3a9276f381d15521e880d7c0bfb7040807af1cac [formerly 9b4cdc482af6f7d8182880b8ef9e0a775ac33546]
Former-commit-id: 1ef4bafbd60a2f594b0737fadf58ec636978b278