Simplify tuple command line arguments error handling
Reuse the boost::program_options::error class (which we already catch in
main()) instead of throwing a std::string, and simplify the throwing
logic accordingly.

There is no need to provide incredibly verbose error messages since the
user is expected to be able to read --help or the man page to get here
in the first place.


Former-commit-id: 3cc0cb20811c8c74864f3a51189d06075a4d13e4
Former-commit-id: 503e68769d0f2d54d5890f735d7e035cf3fa483e
Former-commit-id: 737bee1a2f66c37cf1fd3a55012b1a3803c98095
Former-commit-id: cd765a0138f1a78097340f7dff78f4ef7bd7fce5 [formerly 67a197d9eff7b09b9d690a003724a01efdb2af26]
Former-commit-id: fd32ee3e650b9fa04376cd80926312fff89c7826