revert PR 98
This commit removes the "carryover_wml" feature which was added
just before 1.12 was branched off.

I added this feature at a time when I didn't have as good an
understanding as I do now of the networked mp code, I now believe
that it was poorly concieved and should not be implemented this
way if at all. There are better ways to achieve what was meant to
be achieved here, we are better off to remove it as it currently
exists.


Former-commit-id: 519da202882d7f182ae335e8ff0ad4ca6a7ae048
Former-commit-id: c6fc7e5d13143630283df71b218ce86ea4cba1bb
Former-commit-id: ee257696bf865d9f8b29549a1af9b6b826fc8bc3
Former-commit-id: 8017bf54eb98e19aa377ba826012b3bca0e2e8ab [formerly 4a35c20017b21f6383789a82971f565b97dc9cb4]
Former-commit-id: a52dc8a6d9d0c18dde4f2b532bc44ef26055287d