revert PR 98
This is a parallel commit to PR 243, on master instead of 1.12
branch.

The commit is significantly different because this feature got
swept up in a different refactor of playcampaign / carryover. I
didn't test the feature extensively on this branch, this might not
be necessary at all. I'm simply making the sister commit anyways
so that we have something concrete to discuss when deciding
whether to merge either or both of these pull requests.

I do think the feature is of questionable value and adds
complexity to an already confusing part of the engine, and should
therefore be considered for removal.


Former-commit-id: 9cef7d8720bdfeeaa7332e53072387a0b6efdf8a
Former-commit-id: 775b2653fb76c50d3ee159464944897358605f8d
Former-commit-id: 2e08419474e4d18a54f21b82fa1dc8bfde2666ea
Former-commit-id: ed6ad12781b7296342b4c2144476a7b107fdc4b6 [formerly 9c476511007f742a9bd62fa38cf7793696861ef8]
Former-commit-id: 7d4b691041d11aae914bd9744f96503456be3069